@@ -1,8 +1,8 @@
 apply plugin: 'com.android.application'
 
 ext {
-    supportLibraryVersion = "23.0.1"
-    wearableVersion = "2.0.1"    
+    wearableVersion = "2.0.1"
+    playServicesWearable = "16.0.1"
 }
 
 def generateGitBuild = { ->
@@ -27,12 +27,11 @@ def generateGitBuild = { ->
 }
 
 android {
-    compileSdkVersion 23
-    buildToolsVersion "26.0.2"
+    compileSdkVersion 27
 
     defaultConfig {
         applicationId "info.nightscout.androidaps"
-        minSdkVersion 20
+        minSdkVersion 23
         targetSdkVersion 23
         versionCode 1
         versionName "1.0.2"
@@ -45,8 +44,38 @@ android {
         }
     }
 
-    publishNonDefault true
+    compileOptions {
+        sourceCompatibility JavaVersion.VERSION_1_8
+        targetCompatibility JavaVersion.VERSION_1_8
+    }
 
+    lintOptions {
+        abortOnError false
+    }
+
+    productFlavors {
+        flavorDimensions "standard"
+        full {
+            applicationId "info.nightscout.androidaps"
+            dimension "standard"
+            versionName version
+        }
+        pumpcontrol {
+            applicationId "info.nightscout.aapspumpcontrol"
+            dimension "standard"
+            versionName version + "-pumpcontrol"
+        }
+        nsclient {
+            applicationId "info.nightscout.nsclient"
+            dimension "standard"
+            versionName version + "-nsclient"
+        }
+        nsclient2 {
+            applicationId "info.nightscout.nsclient2"
+            dimension "standard"
+            versionName version + "-nsclient"
+        }
+    }
 }
 
 allprojects {
@@ -60,13 +89,14 @@ allprojects {
 
 
 dependencies {
-    compile fileTree(include: ['*.jar'], dir: 'libs')
-    compile files("libs/hellocharts-library-1.5.5.jar")
+    implementation fileTree(include: ['*.jar'], dir: 'libs')
+    implementation files("libs/hellocharts-library-1.5.5.jar")
     //compile "com.ustwo.android:clockwise-wearable:1.0.2"
-    provided "com.google.android.wearable:wearable:${wearableVersion}"
-    compile "com.google.android.support:wearable:${wearableVersion}"
-    compile "com.google.android.gms:play-services-wearable:7.3.0"
-    compile(name:"ustwo-clockwise-debug", ext:"aar")
-    compile "com.android.support:support-v4:23.0.1"
-    compile "me.denley.wearpreferenceactivity:wearpreferenceactivity:0.5.0"
+    compileOnly "com.google.android.wearable:wearable:${wearableVersion}"
+    implementation "com.google.android.support:wearable:${wearableVersion}"
+    implementation "com.google.android.gms:play-services-wearable:${playServicesWearable}"
+    implementation(name:"ustwo-clockwise-debug", ext:"aar")
+    implementation "com.android.support:support-v4:27.0.1"
+    implementation 'com.android.support:wear:27.0.1'
+    implementation "me.denley.wearpreferenceactivity:wearpreferenceactivity:0.5.0"
 }
