@@ -1,32 +1,35 @@
 package info.nightscout.androidaps;
 
-import android.Manifest;
-import android.app.Activity;
-import android.content.ActivityNotFoundException;
 import android.content.Context;
-import android.content.DialogInterface;
 import android.content.Intent;
 import android.content.pm.PackageManager;
 import android.graphics.Rect;
-import android.net.Uri;
-import android.os.Build;
 import android.os.Bundle;
+import android.os.PersistableBundle;
 import android.os.PowerManager;
-import android.provider.Settings;
+import android.support.annotation.Nullable;
+import android.support.design.widget.NavigationView;
+import android.support.design.widget.TabLayout;
 import android.support.v4.app.ActivityCompat;
-import android.support.v4.content.ContextCompat;
 import android.support.v4.view.ViewPager;
+import android.support.v4.widget.DrawerLayout;
+import android.support.v7.app.ActionBarDrawerToggle;
 import android.support.v7.app.AlertDialog;
 import android.support.v7.app.AppCompatActivity;
-import android.support.v7.widget.PopupMenu;
-import android.view.MenuInflater;
+import android.support.v7.widget.Toolbar;
+import android.text.SpannableString;
+import android.text.method.LinkMovementMethod;
+import android.text.util.Linkify;
+import android.util.TypedValue;
+import android.view.Menu;
 import android.view.MenuItem;
 import android.view.MotionEvent;
 import android.view.View;
 import android.view.WindowManager;
 import android.view.inputmethod.InputMethodManager;
 import android.widget.EditText;
-import android.widget.ImageButton;
+import android.widget.LinearLayout;
+import android.widget.TextView;
 
 import com.joanzapata.iconify.Iconify;
 import com.joanzapata.iconify.fonts.FontAwesomeModule;
@@ -35,113 +38,229 @@ import com.squareup.otto.Subscribe;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import info.nightscout.androidaps.Services.AlarmSoundService;
+import info.nightscout.androidaps.activities.AgreementActivity;
+import info.nightscout.androidaps.activities.HistoryBrowseActivity;
+import info.nightscout.androidaps.activities.PreferencesActivity;
+import info.nightscout.androidaps.activities.SingleFragmentActivity;
 import info.nightscout.androidaps.data.Profile;
 import info.nightscout.androidaps.events.EventAppExit;
+import info.nightscout.androidaps.events.EventFeatureRunning;
 import info.nightscout.androidaps.events.EventPreferenceChange;
 import info.nightscout.androidaps.events.EventRefreshGui;
 import info.nightscout.androidaps.interfaces.PluginBase;
-import info.nightscout.androidaps.plugins.ConfigBuilder.ConfigBuilderPlugin;
-import info.nightscout.androidaps.plugins.Overview.events.EventSetWakeLock;
-import info.nightscout.androidaps.tabs.SlidingTabLayout;
+import info.nightscout.androidaps.interfaces.PluginType;
+import info.nightscout.androidaps.logging.L;
+import info.nightscout.androidaps.plugins.aps.loop.LoopPlugin;
+import info.nightscout.androidaps.plugins.configBuilder.ProfileFunctions;
+import info.nightscout.androidaps.plugins.general.nsclient.data.NSSettingsStatus;
+import info.nightscout.androidaps.plugins.general.versionChecker.VersionCheckerUtilsKt;
+import info.nightscout.androidaps.setupwizard.SetupWizardActivity;
 import info.nightscout.androidaps.tabs.TabPageAdapter;
-import info.nightscout.utils.ImportExportPrefs;
-import info.nightscout.utils.LocaleHelper;
-import info.nightscout.utils.LogDialog;
-import info.nightscout.utils.OKDialog;
-import info.nightscout.utils.PasswordProtection;
-import info.nightscout.utils.SP;
-import info.nightscout.utils.ToastUtils;
+import info.nightscout.androidaps.utils.AndroidPermission;
+import info.nightscout.androidaps.utils.FabricPrivacy;
+import info.nightscout.androidaps.utils.LocaleHelper;
+import info.nightscout.androidaps.utils.OKDialog;
+import info.nightscout.androidaps.utils.PasswordProtection;
+import info.nightscout.androidaps.utils.SP;
 
-public class MainActivity extends AppCompatActivity implements View.OnClickListener {
-    private static Logger log = LoggerFactory.getLogger(MainActivity.class);
-
-    static final int CASE_STORAGE = 0x1;
-    static final int CASE_SMS = 0x2;
-    static final int CASE_LOCATION = 0x3;
-
-    private boolean askForSMS = false;
-    private boolean askForLocation = true;
-
-    ImageButton menuButton;
+public class MainActivity extends AppCompatActivity {
+    private static Logger log = LoggerFactory.getLogger(L.CORE);
 
     protected PowerManager.WakeLock mWakeLock;
+
+    private ActionBarDrawerToggle actionBarDrawerToggle;
+
+    private MenuItem pluginPreferencesMenuItem;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        Iconify.with(new FontAwesomeModule());
-        LocaleHelper.onCreate(this, "en");
-        setContentView(R.layout.activity_main);
-        menuButton = (ImageButton) findViewById(R.id.overview_menuButton);
-        menuButton.setOnClickListener(this);
 
-        checkEula();
-        if (Build.VERSION.SDK_INT > Build.VERSION_CODES.LOLLIPOP_MR1) {
-            askForPermission(new String[]{Manifest.permission.READ_EXTERNAL_STORAGE,
-                    Manifest.permission.WRITE_EXTERNAL_STORAGE}, CASE_STORAGE);
-        }
-        askForBatteryOptimizationPermission();
-        doMigrations();
-        if (Config.logFunctionCalls)
+        if (L.isEnabled(L.CORE))
             log.debug("onCreate");
 
-        onStatusEvent(new EventSetWakeLock(SP.getBoolean("lockscreen", false)));
+        Iconify.with(new FontAwesomeModule());
+        LocaleHelper.onCreate(this, "en");
+
+        setContentView(R.layout.activity_main);
+        setSupportActionBar(findViewById(R.id.toolbar));
+        getSupportActionBar().setDisplayShowTitleEnabled(false);
+        getSupportActionBar().setDisplayHomeAsUpEnabled(true);
+        getSupportActionBar().setHomeButtonEnabled(true);
+
+        DrawerLayout drawerLayout = findViewById(R.id.drawer_layout);
+        actionBarDrawerToggle = new ActionBarDrawerToggle(this, drawerLayout, R.string.open_navigation, R.string.close_navigation);
+        drawerLayout.addDrawerListener(actionBarDrawerToggle);
+        actionBarDrawerToggle.syncState();
+
+        // initialize screen wake lock
+        onEventPreferenceChange(new EventPreferenceChange(R.string.key_keep_screen_on));
+
+        doMigrations();
 
         registerBus();
-        setUpTabs(false);
+        setupTabs();
+        setupViews(false);
+
+        final ViewPager viewPager = findViewById(R.id.pager);
+        viewPager.addOnPageChangeListener(new ViewPager.OnPageChangeListener() {
+            @Override
+            public void onPageScrolled(int position, float positionOffset, int positionOffsetPixels) {
+            }
+
+            @Override
+            public void onPageSelected(int position) {
+                checkPluginPreferences(viewPager);
+            }
+
+            @Override
+            public void onPageScrollStateChanged(int state) {
+            }
+        });
+
+        //Check here if loop plugin is disabled. Else check via constraints
+        if (!LoopPlugin.getPlugin().isEnabled(PluginType.LOOP))
+            VersionCheckerUtilsKt.triggerCheckVersion();
+
+        FabricPrivacy.setUserStats();
+    }
+
+    private void checkPluginPreferences(ViewPager viewPager) {
+        if (pluginPreferencesMenuItem == null) return;
+        if (((TabPageAdapter) viewPager.getAdapter()).getPluginAt(viewPager.getCurrentItem()).getPreferencesId() != -1)
+            pluginPreferencesMenuItem.setEnabled(true);
+        else pluginPreferencesMenuItem.setEnabled(false);
+    }
+
+    @Override
+    public void onPostCreate(@Nullable Bundle savedInstanceState, @Nullable PersistableBundle persistentState) {
+        super.onPostCreate(savedInstanceState, persistentState);
+        actionBarDrawerToggle.syncState();
+    }
+
+    @Override
+    protected void onResume() {
+        super.onResume();
+
+        if (L.isEnabled(L.CORE))
+            log.debug("onResume");
+
+        if (!SP.getBoolean(R.string.key_setupwizard_processed, false)) {
+            Intent intent = new Intent(this, SetupWizardActivity.class);
+            startActivity(intent);
+        } else {
+            checkEula();
+        }
+
+        AndroidPermission.notifyForStoragePermission(this);
+        AndroidPermission.notifyForBatteryOptimizationPermission(this);
+        if (Config.PUMPDRIVERS) {
+            AndroidPermission.notifyForLocationPermissions(this);
+            AndroidPermission.notifyForSMSPermissions(this);
+        }
+
+        MainApp.bus().post(new EventFeatureRunning(EventFeatureRunning.Feature.MAIN));
+    }
+
+    @Override
+    public void onDestroy() {
+        if (L.isEnabled(L.CORE))
+            log.debug("onDestroy");
+        if (mWakeLock != null)
+            if (mWakeLock.isHeld())
+                mWakeLock.release();
+        super.onDestroy();
     }
 
     @Subscribe
-    public void onStatusEvent(final EventSetWakeLock ev) {
-        final PowerManager pm = (PowerManager) getSystemService(Context.POWER_SERVICE);
-        if (ev.lock) {
-            mWakeLock = pm.newWakeLock(PowerManager.SCREEN_BRIGHT_WAKE_LOCK, "AAPS");
-            if (!mWakeLock.isHeld())
-                mWakeLock.acquire();
-        } else {
-            if (mWakeLock != null && mWakeLock.isHeld())
-                mWakeLock.release();
+    public void onEventPreferenceChange(final EventPreferenceChange ev) {
+        if (ev.isChanged(R.string.key_keep_screen_on)) {
+            boolean keepScreenOn = SP.getBoolean(R.string.key_keep_screen_on, false);
+            final PowerManager pm = (PowerManager) getSystemService(Context.POWER_SERVICE);
+            if (keepScreenOn) {
+                mWakeLock = pm.newWakeLock(PowerManager.SCREEN_BRIGHT_WAKE_LOCK, "AndroidAPS:MainActivity_onEventPreferenceChange");
+                if (!mWakeLock.isHeld())
+                    mWakeLock.acquire();
+            } else {
+                if (mWakeLock != null && mWakeLock.isHeld())
+                    mWakeLock.release();
+            }
         }
     }
 
     @Subscribe
     public void onStatusEvent(final EventRefreshGui ev) {
-        String lang = SP.getString("language", "en");
+        String lang = SP.getString(R.string.key_language, "en");
         LocaleHelper.setLocale(getApplicationContext(), lang);
-        runOnUiThread(new Runnable() {
-            @Override
-            public void run() {
-                if(ev.recreate) {
-                    recreate();
-                }else {
-                    try { // activity may be destroyed
-                        setUpTabs(true);
-                    } catch (IllegalStateException e) {
-                        log.error("Unhandled exception", e);
-                    }
+        runOnUiThread(() -> {
+            if (ev.recreate) {
+                recreate();
+            } else {
+                try { // activity may be destroyed
+                    setupTabs();
+                    setupViews(true);
+                } catch (IllegalStateException e) {
+                    log.error("Unhandled exception", e);
                 }
-
-                boolean lockScreen = BuildConfig.NSCLIENTOLNY && SP.getBoolean("lockscreen", false);
-                if (lockScreen)
-                    getWindow().addFlags(WindowManager.LayoutParams.FLAG_KEEP_SCREEN_ON);
-                else
-                    getWindow().clearFlags(WindowManager.LayoutParams.FLAG_KEEP_SCREEN_ON);
             }
+
+            boolean keepScreenOn = Config.NSCLIENT && SP.getBoolean(R.string.key_keep_screen_on, false);
+            if (keepScreenOn)
+                getWindow().addFlags(WindowManager.LayoutParams.FLAG_KEEP_SCREEN_ON);
+            else
+                getWindow().clearFlags(WindowManager.LayoutParams.FLAG_KEEP_SCREEN_ON);
         });
     }
 
-    private void setUpTabs(boolean switchToLast) {
+    private void setupViews(boolean switchToLast) {
         TabPageAdapter pageAdapter = new TabPageAdapter(getSupportFragmentManager(), this);
+        NavigationView navigationView = findViewById(R.id.navigation_view);
+        navigationView.setNavigationItemSelectedListener(menuItem -> {
+            return true;
+        });
+        Menu menu = navigationView.getMenu();
+        menu.clear();
         for (PluginBase p : MainApp.getPluginsList()) {
             pageAdapter.registerNewFragment(p);
+            if (p.hasFragment() && !p.isFragmentVisible() && p.isEnabled(p.pluginDescription.getType()) && !p.pluginDescription.neverVisible) {
+                MenuItem menuItem = menu.add(p.getName());
+                menuItem.setCheckable(true);
+                menuItem.setOnMenuItemClickListener(item -> {
+                    Intent intent = new Intent(this, SingleFragmentActivity.class);
+                    intent.putExtra("plugin", MainApp.getPluginsList().indexOf(p));
+                    startActivity(intent);
+                    ((DrawerLayout) findViewById(R.id.drawer_layout)).closeDrawers();
+                    return true;
+                });
+            }
         }
-        ViewPager mPager = (ViewPager) findViewById(R.id.pager);
+        ViewPager mPager = findViewById(R.id.pager);
         mPager.setAdapter(pageAdapter);
-        SlidingTabLayout mTabs = (SlidingTabLayout) findViewById(R.id.tabs);
-        mTabs.setViewPager(mPager);
         if (switchToLast)
             mPager.setCurrentItem(pageAdapter.getCount() - 1, false);
+        checkPluginPreferences(mPager);
+    }
+
+    private void setupTabs() {
+        ViewPager viewPager = findViewById(R.id.pager);
+        TabLayout normalTabs = findViewById(R.id.tabs_normal);
+        normalTabs.setupWithViewPager(viewPager, true);
+        TabLayout compactTabs = findViewById(R.id.tabs_compact);
+        compactTabs.setupWithViewPager(viewPager, true);
+        Toolbar toolbar = findViewById(R.id.toolbar);
+        if (SP.getBoolean("short_tabtitles", false)) {
+            normalTabs.setVisibility(View.GONE);
+            compactTabs.setVisibility(View.VISIBLE);
+            toolbar.setLayoutParams(new LinearLayout.LayoutParams(Toolbar.LayoutParams.MATCH_PARENT, (int) getResources().getDimension(R.dimen.compact_height)));
+        } else {
+            normalTabs.setVisibility(View.VISIBLE);
+            compactTabs.setVisibility(View.GONE);
+            TypedValue typedValue = new TypedValue();
+            if (getTheme().resolveAttribute(R.attr.actionBarSize, typedValue, true)) {
+                toolbar.setLayoutParams(new LinearLayout.LayoutParams(Toolbar.LayoutParams.MATCH_PARENT,
+                        TypedValue.complexToDimensionPixelSize(typedValue.data, getResources().getDisplayMetrics())));
+            }
+        }
     }
 
     private void registerBus() {
@@ -171,7 +290,7 @@ public class MainActivity extends AppCompatActivity implements View.OnClickListe
         // Added in 1.57 at 21.01.2018
         Integer unreachable_threshold = SP.getInt(R.string.key_pump_unreachable_threshold, 30);
         SP.remove(R.string.key_pump_unreachable_threshold);
-        if(unreachable_threshold < 30) unreachable_threshold = 30;
+        if (unreachable_threshold < 30) unreachable_threshold = 30;
         SP.putString(R.string.key_pump_unreachable_threshold, unreachable_threshold.toString());
     }
 
@@ -179,7 +298,7 @@ public class MainActivity extends AppCompatActivity implements View.OnClickListe
     private void checkUpgradeToProfileTarget() { // TODO: can be removed in the future
         boolean oldKeyExists = SP.contains("openapsma_min_bg");
         if (oldKeyExists) {
-            Profile profile = MainApp.getConfigBuilder().getProfile();
+            Profile profile = ProfileFunctions.getInstance().getProfile();
             String oldRange = SP.getDouble("openapsma_min_bg", 0d) + " - " + SP.getDouble("openapsma_max_bg", 0d);
             String newRange = "";
             if (profile != null) {
@@ -199,119 +318,23 @@ public class MainActivity extends AppCompatActivity implements View.OnClickListe
         }
     }
 
-    //check for sms permission if enable in prefernces
-    @Subscribe
-    public void onStatusEvent(final EventPreferenceChange ev) {
-        if (ev.isChanged(R.string.key_smscommunicator_remotecommandsallowed)) {
-            if (Build.VERSION.SDK_INT > Build.VERSION_CODES.LOLLIPOP_MR1) {
-                synchronized (this) {
-                    if (SP.getBoolean(R.string.key_smscommunicator_remotecommandsallowed, false)) {
-                        setAskForSMS();
-                    }
-                }
-            }
-        }
-    }
-
-    private synchronized void setAskForSMS() {
-        askForSMS = true;
-    }
-
-    @Override
-    protected void onResume() {
-        super.onResume();
-        askForSMSPermissions();
-        askForLocationPermissions();
-    }
-
-    @Override
-    public void onDestroy() {
-        if (mWakeLock != null)
-            if (mWakeLock.isHeld())
-                mWakeLock.release();
-        super.onDestroy();
-    }
-
-    private void askForBatteryOptimizationPermission() {
-        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.M) {
-            final String packageName = getPackageName();
-
-            final PowerManager pm = (PowerManager) getSystemService(Context.POWER_SERVICE);
-            if (!pm.isIgnoringBatteryOptimizations(packageName)) {
-                log.debug("Requesting ignore battery optimization");
-
-                OKDialog.show(this, getString(R.string.pleaseallowpermission), String.format(getString(R.string.needwhitelisting), getString(R.string.app_name)), new Runnable() {
-
-                    @Override
-                    public void run() {
-                        try {
-                            final Intent intent = new Intent();
-
-                            // ignoring battery optimizations required for constant connection
-                            intent.setAction(Settings.ACTION_REQUEST_IGNORE_BATTERY_OPTIMIZATIONS);
-                            intent.setData(Uri.parse("package:" + packageName));
-                            startActivity(intent);
-
-                        } catch (ActivityNotFoundException e) {
-                            final String msg = getString(R.string.batteryoptimalizationerror);
-                            ToastUtils.showToastInUiThread(getApplicationContext(), msg);
-                            log.error(msg);
-                        }
-                    }
-                });
-            }
-        }
-    }
-
-    private synchronized void askForSMSPermissions() {
-        if (askForSMS) { //only when settings were changed an MainActivity resumes.
-            askForSMS = false;
-            if (SP.getBoolean(R.string.smscommunicator_remotecommandsallowed, false)) {
-                if (Build.VERSION.SDK_INT > Build.VERSION_CODES.LOLLIPOP_MR1) {
-                    askForPermission(new String[]{Manifest.permission.RECEIVE_SMS,
-                            Manifest.permission.SEND_SMS,
-                            Manifest.permission.RECEIVE_MMS}, CASE_SMS);
-                }
-            }
-        }
-    }
-
-    private synchronized void askForLocationPermissions() {
-        if (askForLocation) { //only when settings were changed an MainActivity resumes.
-            askForLocation = false;
-            if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.M) {
-                askForPermission(new String[]{Manifest.permission.ACCESS_FINE_LOCATION,
-                        Manifest.permission.ACCESS_FINE_LOCATION,
-                        Manifest.permission.ACCESS_FINE_LOCATION}, CASE_LOCATION);
-            }
-        }
-    }
-
-    private void askForPermission(String[] permission, Integer requestCode) {
-        boolean test = false;
-        for (int i = 0; i < permission.length; i++) {
-            test = test || (ContextCompat.checkSelfPermission(this, permission[i]) != PackageManager.PERMISSION_GRANTED);
-        }
-        if (test) {
-            ActivityCompat.requestPermissions(this, permission, requestCode);
-        }
-    }
-
     @Override
     public void onRequestPermissionsResult(int requestCode, String[] permissions, int[] grantResults) {
         super.onRequestPermissionsResult(requestCode, permissions, grantResults);
         if (permissions.length != 0) {
             if (ActivityCompat.checkSelfPermission(this, permissions[0]) == PackageManager.PERMISSION_GRANTED) {
                 switch (requestCode) {
-                    case CASE_STORAGE:
+                    case AndroidPermission.CASE_STORAGE:
                         //show dialog after permission is granted
                         AlertDialog.Builder alert = new AlertDialog.Builder(this);
                         alert.setMessage(R.string.alert_dialog_storage_permission_text);
                         alert.setPositiveButton(R.string.ok, null);
                         alert.show();
                         break;
-                    case CASE_LOCATION:
-                    case CASE_SMS:
+                    case AndroidPermission.CASE_LOCATION:
+                    case AndroidPermission.CASE_SMS:
+                    case AndroidPermission.CASE_BATTERY:
+                    case AndroidPermission.CASE_PHONESTATE:
                         break;
                 }
             }
@@ -336,82 +359,67 @@ public class MainActivity extends AppCompatActivity implements View.OnClickListe
     }
 
     @Override
-    public void onClick(final View v) {
-        final Activity activity = this;
-        switch (v.getId()) {
-            case R.id.overview_menuButton:
-                PopupMenu popup = new PopupMenu(v.getContext(), v);
-                MenuInflater inflater = popup.getMenuInflater();
-                inflater.inflate(R.menu.menu_main, popup.getMenu());
-                popup.setOnMenuItemClickListener(new PopupMenu.OnMenuItemClickListener() {
-                    @Override
-                    public boolean onMenuItemClick(MenuItem item) {
-                        int id = item.getItemId();
-                        switch (id) {
-                            case R.id.nav_preferences:
-                                PasswordProtection.QueryPassword(v.getContext(), R.string.settings_password, "settings_password", new Runnable() {
-                                    @Override
-                                    public void run() {
-                                        Intent i = new Intent(v.getContext(), PreferencesActivity.class);
-                                        i.putExtra("id", -1);
-                                        startActivity(i);
-                                    }
-                                }, null);
-                                break;
-                            case R.id.nav_resetdb:
-                                new AlertDialog.Builder(v.getContext())
-                                        .setTitle(R.string.nav_resetdb)
-                                        .setMessage(R.string.reset_db_confirm)
-                                        .setNegativeButton(android.R.string.cancel, null)
-                                        .setPositiveButton(android.R.string.ok, new DialogInterface.OnClickListener() {
-                                            @Override
-                                            public void onClick(DialogInterface dialog, int which) {
-                                                MainApp.getDbHelper().resetDatabases();
-                                            }
-                                        })
-                                        .create()
-                                        .show();
-                                break;
-                            case R.id.nav_export:
-                                ImportExportPrefs.verifyStoragePermissions(activity);
-                                ImportExportPrefs.exportSharedPreferences(activity);
-                                break;
-                            case R.id.nav_import:
-                                ImportExportPrefs.verifyStoragePermissions(activity);
-                                ImportExportPrefs.importSharedPreferences(activity);
-                                break;
-                            case R.id.nav_show_logcat:
-                                LogDialog.showLogcat(v.getContext());
-                                break;
-                            case R.id.nav_about:
-                                AlertDialog.Builder builder = new AlertDialog.Builder(v.getContext());
-                                builder.setTitle(getString(R.string.app_name) + " " + BuildConfig.VERSION);
-                                if (Config.NSCLIENT|| Config.G5UPLOADER)
-                                    builder.setIcon(R.mipmap.yellowowl);
-                                else
-                                    builder.setIcon(R.mipmap.blueowl);
-                                String message = "Build: " + BuildConfig.BUILDVERSION + "\n";
-                                message += MainApp.sResources.getString(R.string.configbuilder_nightscoutversion_label) + " " + ConfigBuilderPlugin.nightscoutVersionName;
-                                builder.setMessage(message);
-                                builder.setPositiveButton(MainApp.sResources.getString(R.string.ok), null);
-                                AlertDialog alertDialog = builder.create();
-                                alertDialog.show();
-                                break;
-                            case R.id.nav_exit:
-                                log.debug("Exiting");
-                                MainApp.instance().stopKeepAliveService();
-                                MainApp.bus().post(new EventAppExit());
-                                MainApp.closeDbHelper();
-                                finish();
-                                System.runFinalization();
-                                System.exit(0);
-                                break;
-                        }
-                        return false;
-                    }
-                });
-                popup.show();
-                break;
+    public boolean onCreateOptionsMenu(Menu menu) {
+        getMenuInflater().inflate(R.menu.menu_main, menu);
+        pluginPreferencesMenuItem = menu.findItem(R.id.nav_plugin_preferences);
+        checkPluginPreferences(findViewById(R.id.pager));
+        return true;
+    }
+
+    @Override
+    public boolean onOptionsItemSelected(MenuItem item) {
+        int id = item.getItemId();
+        switch (id) {
+            case R.id.nav_preferences:
+                PasswordProtection.QueryPassword(this, R.string.settings_password, "settings_password", () -> {
+                    Intent i = new Intent(this, PreferencesActivity.class);
+                    i.putExtra("id", -1);
+                    startActivity(i);
+                }, null);
+                return true;
+            case R.id.nav_historybrowser:
+                startActivity(new Intent(this, HistoryBrowseActivity.class));
+                return true;
+            case R.id.nav_setupwizard:
+                startActivity(new Intent(this, SetupWizardActivity.class));
+                return true;
+            case R.id.nav_about:
+                AlertDialog.Builder builder = new AlertDialog.Builder(this);
+                builder.setTitle(MainApp.gs(R.string.app_name) + " " + BuildConfig.VERSION);
+                builder.setIcon(MainApp.getIcon());
+                String message = "Build: " + BuildConfig.BUILDVERSION + "\n";
+                message += "Flavor: " + BuildConfig.FLAVOR + BuildConfig.BUILD_TYPE + "\n";
+                message += MainApp.gs(R.string.configbuilder_nightscoutversion_label) + " " + NSSettingsStatus.getInstance().nightscoutVersionName;
+                if (MainApp.engineeringMode)
+                    message += "\n" + MainApp.gs(R.string.engineering_mode_enabled);
+                message += MainApp.gs(R.string.about_link_urls);
+                final SpannableString messageSpanned = new SpannableString(message);
+                Linkify.addLinks(messageSpanned, Linkify.WEB_URLS);
+                builder.setMessage(messageSpanned);
+                builder.setPositiveButton(MainApp.gs(R.string.ok), null);
+                AlertDialog alertDialog = builder.create();
+                alertDialog.show();
+                ((TextView) alertDialog.findViewById(android.R.id.message)).setMovementMethod(LinkMovementMethod.getInstance());
+                return true;
+            case R.id.nav_exit:
+                log.debug("Exiting");
+                MainApp.instance().stopKeepAliveService();
+                MainApp.bus().post(new EventAppExit());
+                MainApp.closeDbHelper();
+                finish();
+                System.runFinalization();
+                System.exit(0);
+                return true;
+            case R.id.nav_plugin_preferences:
+                ViewPager viewPager = findViewById(R.id.pager);
+                final PluginBase plugin = ((TabPageAdapter) viewPager.getAdapter()).getPluginAt(viewPager.getCurrentItem());
+                PasswordProtection.QueryPassword(this, R.string.settings_password, "settings_password", () -> {
+                    Intent i = new Intent(this, PreferencesActivity.class);
+                    i.putExtra("id", plugin.getPreferencesId());
+                    startActivity(i);
+                }, null);
+                return true;
         }
+        return actionBarDrawerToggle.onOptionsItemSelected(item);
     }
 }
