@@ -5,20 +5,23 @@ import org.slf4j.LoggerFactory;
 
 import java.lang.reflect.Field;
 import java.lang.reflect.Method;
+import java.util.ConcurrentModificationException;
 import java.util.HashMap;
 import java.util.HashSet;
 import java.util.Map;
 import java.util.Set;
 
 import info.nightscout.androidaps.events.Event;
+import info.nightscout.androidaps.logging.L;
 
-/** Logs events has they're being posted to and dispatched from the event bus.
- *
+/**
+ * Logs events has they're being posted to and dispatched from the event bus.
+ * <p>
  * A summary of event-receiver calls that occurred so far is logged
  * after 10s (after startup) and then again every 60s.
- * */
+ */
 public class LoggingBus extends Bus {
-    private static Logger log = LoggerFactory.getLogger(LoggingBus.class);
+    private static Logger log = LoggerFactory.getLogger(L.EVENTS);
 
     private static long everyMinute = System.currentTimeMillis() + 10 * 1000;
     private Map<String, Set<String>> event2Receiver = new HashMap<>();
@@ -48,7 +51,10 @@ public class LoggingBus extends Bus {
             log.debug("    source: <unknown>");
         }
 
-        super.post(event);
+        try {
+            super.post(event);
+        } catch (IllegalStateException ignored) {
+        }
     }
 
     @Override
@@ -70,16 +76,19 @@ public class LoggingBus extends Bus {
             log.debug("    receiver: <unknown>");
         }
 
-        if (everyMinute < System.currentTimeMillis()) {
-            log.debug("***************** Event -> receiver pairings seen so far ****************");
-            for (Map.Entry<String, Set<String>> stringSetEntry : event2Receiver.entrySet()) {
-                log.debug("  " + stringSetEntry.getKey());
-                for (String s : stringSetEntry.getValue()) {
-                    log.debug("    -> " + s);
+        try {
+            if (everyMinute < System.currentTimeMillis()) {
+                log.debug("***************** Event -> receiver pairings seen so far ****************");
+                for (Map.Entry<String, Set<String>> stringSetEntry : event2Receiver.entrySet()) {
+                    log.debug("  " + stringSetEntry.getKey());
+                    for (String s : stringSetEntry.getValue()) {
+                        log.debug("    -> " + s);
+                    }
                 }
+                log.debug("*************************************************************************");
+                everyMinute = System.currentTimeMillis() + 60 * 1000;
             }
-            log.debug("*************************************************************************");
-            everyMinute = System.currentTimeMillis() + 60 * 1000;
+        } catch (ConcurrentModificationException ignored) {
         }
 
         super.dispatch(event, wrapper);
