@@ -6,10 +6,8 @@ import org.json.JSONObject;
 
 import java.util.Date;
 
-import info.nightscout.androidaps.MainApp;
 import info.nightscout.androidaps.db.CareportalEvent;
 import info.nightscout.androidaps.db.Source;
-import info.nightscout.androidaps.interfaces.InsulinInterface;
 
 /**
  * Created by mike on 29.05.2017.
@@ -17,6 +15,7 @@ import info.nightscout.androidaps.interfaces.InsulinInterface;
 
 public class DetailedBolusInfo {
     public long date = System.currentTimeMillis();
+    public long lastKnownBolusTime;
     public String eventType = CareportalEvent.MEALBOLUS;
     public double insulin = 0;
     public double carbs = 0;
@@ -29,14 +28,38 @@ public class DetailedBolusInfo {
     public Context context = null;         // context for progress dialog
     public long pumpId = 0;                // id of record if comming from pump history (not a newly created treatment)
     public boolean isSMB = false;          // is a Super-MicroBolus
+    public long deliverAt = 0;             // SMB should be delivered within 1 min from this time
+    public String notes = null;
+
+    public DetailedBolusInfo copy() {
+        DetailedBolusInfo n = new DetailedBolusInfo();
+        n.date = date;
+        n.eventType = eventType;
+        n.insulin = insulin;
+        n.carbs = carbs;
+        n.source = source;
+        n.isValid = isValid;
+        n.glucose = glucose;
+        n.glucoseType = glucoseType;
+        n.carbTime = carbTime;
+        n.boluscalc = boluscalc;
+        n.context = context;
+        n.pumpId = pumpId;
+        n.isSMB = isSMB;
+        n.deliverAt = deliverAt;
+        n.notes = notes;
+        return n;
+    }
 
     @Override
     public String toString() {
         return new Date(date).toLocaleString() +
+                " date: " + date +
                 " insulin: " + insulin +
                 " carbs: " + carbs +
                 " isValid: " + isValid +
                 " carbTime: " + carbTime +
-                " isSMB: " + isSMB;
+                " isSMB: " + isSMB +
+                " deliverAt: " + new Date(deliverAt).toLocaleString();
     }
 }
