@@ -1,5 +1,7 @@
 package info.nightscout.androidaps.data;
 
+import com.rits.cloning.Cloner;
+
 import org.json.JSONException;
 import org.json.JSONObject;
 import org.slf4j.Logger;
@@ -7,8 +9,8 @@ import org.slf4j.LoggerFactory;
 
 import java.util.Date;
 
-import info.nightscout.utils.DateUtil;
-import info.nightscout.utils.Round;
+import info.nightscout.androidaps.utils.DateUtil;
+import info.nightscout.androidaps.utils.Round;
 
 public class IobTotal {
     private static Logger log = LoggerFactory.getLogger(IobTotal.class);
@@ -21,8 +23,8 @@ public class IobTotal {
     public double hightempinsulin;
 
     // oref1
-    public double microBolusInsulin;
-    public double microBolusIOB;
+    public long lastBolusTime;
+    public IobTotal iobWithZeroTemp;
 
     public double netInsulin = 0d; // for calculations from temp basals only
     public double netRatio = 0d; // net ratio at start of temp basal
@@ -31,6 +33,12 @@ public class IobTotal {
 
     long time;
 
+
+    public IobTotal copy() {
+        Cloner cloner = new Cloner();
+        return cloner.deepClone(this);
+    }
+
     public IobTotal(long time) {
         this.iob = 0d;
         this.activity = 0d;
@@ -38,8 +46,7 @@ public class IobTotal {
         this.basaliob = 0d;
         this.netbasalinsulin = 0d;
         this.hightempinsulin = 0d;
-        this.microBolusInsulin = 0d;
-        this.microBolusIOB = 0d;
+        this.lastBolusTime = 0;
         this.time = time;
     }
 
@@ -52,8 +59,6 @@ public class IobTotal {
         hightempinsulin += other.hightempinsulin;
         netInsulin += other.netInsulin;
         extendedBolusInsulin += other.extendedBolusInsulin;
-        microBolusInsulin += other.microBolusInsulin;
-        microBolusIOB += other.microBolusIOB;
         return this;
     }
 
@@ -62,11 +67,13 @@ public class IobTotal {
         result.iob = bolusIOB.iob + basalIob.basaliob;
         result.activity = bolusIOB.activity + basalIob.activity;
         result.bolussnooze = bolusIOB.bolussnooze;
-        result.basaliob = basalIob.basaliob;
-        result.netbasalinsulin = basalIob.netbasalinsulin;
-        result.hightempinsulin = basalIob.hightempinsulin;
-        result.microBolusInsulin = bolusIOB.microBolusInsulin + basalIob.microBolusInsulin;
-        result.microBolusIOB = bolusIOB.microBolusIOB + basalIob.microBolusIOB;
+        result.basaliob = bolusIOB.basaliob + basalIob.basaliob;
+        result.netbasalinsulin = bolusIOB.netbasalinsulin + basalIob.netbasalinsulin;
+        result.hightempinsulin = basalIob.hightempinsulin + bolusIOB.hightempinsulin;
+        result.netInsulin = basalIob.netInsulin + bolusIOB.netInsulin;
+        result.extendedBolusInsulin = basalIob.extendedBolusInsulin + bolusIOB.extendedBolusInsulin;
+        result.lastBolusTime = bolusIOB.lastBolusTime;
+        result.iobWithZeroTemp = basalIob.iobWithZeroTemp;
         return result;
     }
 
@@ -77,8 +84,8 @@ public class IobTotal {
         this.basaliob = Round.roundTo(this.basaliob, 0.001);
         this.netbasalinsulin = Round.roundTo(this.netbasalinsulin, 0.001);
         this.hightempinsulin = Round.roundTo(this.hightempinsulin, 0.001);
-        this.microBolusInsulin = Round.roundTo(this.microBolusInsulin, 0.001);
-        this.microBolusIOB = Round.roundTo(this.microBolusIOB, 0.001);
+        this.netInsulin = Round.roundTo(this.netInsulin, 0.001);
+        this.extendedBolusInsulin = Round.roundTo(this.extendedBolusInsulin, 0.001);
         return this;
     }
 
@@ -102,7 +109,24 @@ public class IobTotal {
             json.put("basaliob", basaliob);
             json.put("bolussnooze", bolussnooze);
             json.put("activity", activity);
+            json.put("lastBolusTime", lastBolusTime);
             json.put("time", DateUtil.toISOString(new Date(time)));
+            /*
+
+            This is requested by SMB determine_basal but by based on Scott's info
+            it's MDT specific safety check only
+            It's causing rounding issues in determine_basal
+
+            JSONObject lastTemp = new JSONObject();
+            lastTemp.put("date", lastTempDate);
+            lastTemp.put("rate", lastTempRate);
+            lastTemp.put("duration", lastTempDuration);
+            json.put("lastTemp", lastTemp);
+            */
+            if (iobWithZeroTemp != null) {
+                JSONObject iwzt = iobWithZeroTemp.determineBasalJson();
+                json.put("iobWithZeroTemp", iwzt);
+            }
         } catch (JSONException e) {
             log.error("Unhandled exception", e);
         }
