@@ -11,18 +11,26 @@ import info.nightscout.androidaps.interfaces.Interval;
 
 public class OverlappingIntervals<T extends Interval> extends Intervals<T> {
 
+    public OverlappingIntervals() {
+        super();
+    }
+
+    public OverlappingIntervals(Intervals<T> other) {
+        rawData = other.rawData.clone();
+    }
+
     protected synchronized void merge() {
         boolean needToCut = false;
         long cutTime = 0;
 
-        for (int index = rawData.size()-1; index >= 0; index--) { //begin with newest
+        for (int index = rawData.size() - 1; index >= 0; index--) { //begin with newest
             Interval cur = rawData.valueAt(index);
-            if (cur.isEndingEvent()){
+            if (cur.isEndingEvent()) {
                 needToCut = true;
                 cutTime = cur.start();
             } else {
                 //event that is no EndingEvent might need to be stopped by an ending event
-                if(needToCut&&cur.end() > cutTime){
+                if (needToCut && cur.end() > cutTime) {
                     cur.cutEndTo(cutTime);
                 }
             }
@@ -31,9 +39,9 @@ public class OverlappingIntervals<T extends Interval> extends Intervals<T> {
 
     @Nullable
     public synchronized T getValueByInterval(long time) {
-        for (int index = rawData.size()-1; index >= 0; index--) { //begin with newest
+        for (int index = rawData.size() - 1; index >= 0; index--) { //begin with newest
             T cur = rawData.valueAt(index);
-            if (cur.match(time)){
+            if (cur.match(time)) {
                 return cur;
             }
         }
