@@ -3,6 +3,9 @@ package info.nightscout.androidaps.data;
 import android.support.annotation.Nullable;
 import android.support.v4.util.LongSparseArray;
 
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
 import java.util.ArrayList;
 import java.util.List;
 
@@ -16,8 +19,17 @@ import info.nightscout.androidaps.interfaces.Interval;
 // When no interval match the lastest record without duration is used
 
 public class ProfileIntervals<T extends Interval> {
+    private static Logger log = LoggerFactory.getLogger(ProfileIntervals.class);
 
-    private LongSparseArray<T> rawData = new LongSparseArray<>(); // oldest at index 0
+    private LongSparseArray<T> rawData; // oldest at index 0
+
+    public ProfileIntervals () {
+        rawData = new LongSparseArray<>();
+    }
+
+    public ProfileIntervals (ProfileIntervals<T> other) {
+        rawData = other.rawData.clone();
+    }
 
     public synchronized ProfileIntervals reset() {
         rawData = new LongSparseArray<>();
@@ -25,8 +37,10 @@ public class ProfileIntervals<T extends Interval> {
     }
 
     public synchronized void add(T newInterval) {
-        rawData.put(newInterval.start(), newInterval);
-        merge();
+        if (newInterval.isValid()) {
+            rawData.put(newInterval.start(), newInterval);
+            merge();
+        }
     }
 
     public synchronized void add(List<T> list) {
@@ -51,6 +65,13 @@ public class ProfileIntervals<T extends Interval> {
     public synchronized Interval getValueToTime(long time) {
         int index = binarySearch(time);
         if (index >= 0) return rawData.valueAt(index);
+        // if we request data older than first record, use oldest with zero duration instead
+        for (index = 0; index < rawData.size(); index++) {
+            if (rawData.valueAt(index).durationInMsec() == 0) {
+                //log.debug("Requested profile for time: " + DateUtil.dateAndTimeString(time) + ". Providing oldest record: " + rawData.valueAt(0).toString());
+                return rawData.valueAt(index);
+            }
+        }
         return null;
     }
 
@@ -106,5 +127,10 @@ public class ProfileIntervals<T extends Interval> {
 
     public synchronized T getReversed(int index) {
         return rawData.valueAt(size() - 1 - index);
+    }
+
+    @Override
+    public String toString() {
+        return rawData.toString();
     }
 }
