@@ -1,8 +1,5 @@
 package info.nightscout.androidaps.data;
 
-import android.text.Html;
-import android.text.Spanned;
-
 import org.json.JSONException;
 import org.json.JSONObject;
 import org.slf4j.Logger;
@@ -10,30 +7,31 @@ import org.slf4j.LoggerFactory;
 
 import info.nightscout.androidaps.MainApp;
 import info.nightscout.androidaps.R;
-import info.nightscout.utils.DecimalFormatter;
-import info.nightscout.utils.Round;
+import info.nightscout.androidaps.logging.L;
+import info.nightscout.androidaps.utils.DecimalFormatter;
+import info.nightscout.androidaps.utils.Round;
 
-public class PumpEnactResult extends Object {
-    private static Logger log = LoggerFactory.getLogger(PumpEnactResult.class);
+public class PumpEnactResult {
+    private static Logger log = LoggerFactory.getLogger(L.APS);
 
     public boolean success = false;    // request was processed successfully (but possible no change was needed)
     public boolean enacted = false;    // request was processed successfully and change has been made
     public String comment = "";
 
     // Result of basal change
-    public Integer duration = -1;      // duration set [minutes]
-    public Double absolute = -1d;      // absolute rate [U/h] , isPercent = false
-    public Integer percent = -1;       // percent of current basal [%] (100% = current basal), isPercent = true
+    public int duration = -1;      // duration set [minutes]
+    public double absolute = -1d;      // absolute rate [U/h] , isPercent = false
+    public int percent = -1;       // percent of current basal [%] (100% = current basal), isPercent = true
     public boolean isPercent = false;  // if true percent is used, otherwise absolute
     public boolean isTempCancel = false; // if true we are caceling temp basal
     // Result of treatment delivery
-    public Double bolusDelivered = 0d; // real value of delivered insulin
-    public Double carbsDelivered = 0d; // real value of delivered carbs
+    public double bolusDelivered = 0d; // real value of delivered insulin
+    public double carbsDelivered = 0d; // real value of delivered carbs
 
     public boolean queued = false;
 
     public PumpEnactResult success(boolean success) {
-       this.success = success;
+        this.success = success;
         return this;
     }
 
@@ -47,13 +45,18 @@ public class PumpEnactResult extends Object {
         return this;
     }
 
-    public PumpEnactResult duration(Integer duration) {
+    public PumpEnactResult duration(int duration) {
         this.duration = duration;
         return this;
     }
 
-    public PumpEnactResult absolute(Double absolute) {
+    public PumpEnactResult absolute(double absolute) {
         this.absolute = absolute;
+        return this;
+    }
+
+    public PumpEnactResult percent(int percent) {
+        this.percent = percent;
         return this;
     }
 
@@ -67,12 +70,12 @@ public class PumpEnactResult extends Object {
         return this;
     }
 
-    public PumpEnactResult bolusDelivered(Double bolusDelivered) {
+    public PumpEnactResult bolusDelivered(double bolusDelivered) {
         this.bolusDelivered = bolusDelivered;
         return this;
     }
 
-    public PumpEnactResult carbsDelivered(Double carbsDelivered) {
+    public PumpEnactResult carbsDelivered(double carbsDelivered) {
         this.carbsDelivered = carbsDelivered;
         return this;
     }
@@ -82,72 +85,96 @@ public class PumpEnactResult extends Object {
         return this;
     }
 
-     public String log() {
-        return "Success: " + success + " Enacted: " + enacted + " Comment: " + comment + " Duration: " + duration + " Absolute: " + absolute + " Percent: " + percent + " IsPercent: " + isPercent + " Queued: " + queued;
+    public String log() {
+        return "Success: " + success +
+                " Enacted: " + enacted +
+                " Comment: " + comment +
+                " Duration: " + duration +
+                " Absolute: " + absolute +
+                " Percent: " + percent +
+                " IsPercent: " + isPercent +
+                " IsTempCancel: " + isTempCancel +
+                " bolusDelivered: " + bolusDelivered +
+                " carbsDelivered: " + carbsDelivered +
+                " Queued: " + queued;
     }
 
     public String toString() {
-        String ret = MainApp.sResources.getString(R.string.success) + ": " + success;
+        String ret = MainApp.gs(R.string.success) + ": " + success;
         if (enacted) {
-            if (isTempCancel) {
-                ret += "\n" + MainApp.sResources.getString(R.string.enacted) + ": " + enacted;
-                ret += "\n" + MainApp.sResources.getString(R.string.comment) + ": " + comment + "\n" +
-                        MainApp.sResources.getString(R.string.canceltemp);
+            if (bolusDelivered > 0) {
+                ret += "\n" + MainApp.gs(R.string.enacted) + ": " + enacted;
+                ret += "\n" + MainApp.gs(R.string.comment) + ": " + comment;
+                ret += "\n" + MainApp.gs(R.string.configbuilder_insulin)
+                        + ": " + bolusDelivered + " " + MainApp.gs(R.string.insulin_unit_shortname);
+            } else if (isTempCancel) {
+                ret += "\n" + MainApp.gs(R.string.enacted) + ": " + enacted;
+                if (!comment.isEmpty())
+                    ret += "\n" + MainApp.gs(R.string.comment) + ": " + comment;
+                ret += "\n" + MainApp.gs(R.string.canceltemp);
             } else if (isPercent) {
-                ret += "\n" + MainApp.sResources.getString(R.string.enacted) + ": " + enacted;
-                ret += "\n" + MainApp.sResources.getString(R.string.comment) + ": " + comment;
-                ret += "\n" + MainApp.sResources.getString(R.string.duration) + ": " + duration + " min";
-                ret += "\n" + MainApp.sResources.getString(R.string.percent) + ": " + percent + "%";
+                ret += "\n" + MainApp.gs(R.string.enacted) + ": " + enacted;
+                if (!comment.isEmpty())
+                    ret += "\n" + MainApp.gs(R.string.comment) + ": " + comment;
+                ret += "\n" + MainApp.gs(R.string.duration) + ": " + duration + " min";
+                ret += "\n" + MainApp.gs(R.string.percent) + ": " + percent + "%";
             } else {
-                ret += "\n" + MainApp.sResources.getString(R.string.enacted) + ": " + enacted;
-                ret += "\n" + MainApp.sResources.getString(R.string.comment) + ": " + comment;
-                ret += "\n" + MainApp.sResources.getString(R.string.duration) + ": " + duration + " min";
-                ret += "\n" + MainApp.sResources.getString(R.string.absolute) + ": " + absolute + " U/h";
+                ret += "\n" + MainApp.gs(R.string.enacted) + ": " + enacted;
+                if (!comment.isEmpty())
+                    ret += "\n" + MainApp.gs(R.string.comment) + ": " + comment;
+                ret += "\n" + MainApp.gs(R.string.duration) + ": " + duration + " min";
+                ret += "\n" + MainApp.gs(R.string.absolute) + ": " + absolute + " U/h";
             }
         } else {
-            ret += "\n" + MainApp.sResources.getString(R.string.comment) + ": " + comment;
+            ret += "\n" + MainApp.gs(R.string.comment) + ": " + comment;
         }
         return ret;
     }
 
-    public Spanned toSpanned() {
-        String ret = MainApp.sResources.getString(R.string.success) + ": " + success;
+    public String toHtml() {
+        String ret = "<b>" + MainApp.gs(R.string.success) + "</b>: " + success;
         if (queued) {
-            ret = MainApp.sResources.getString(R.string.waitingforpumpresult);
+            ret = MainApp.gs(R.string.waitingforpumpresult);
         } else if (enacted) {
-            if (isTempCancel) {
-                ret += "<br><b>" + MainApp.sResources.getString(R.string.enacted) + "</b>: " + enacted;
-                ret += "<br><b>" + MainApp.sResources.getString(R.string.comment) + "</b>: " + comment +
-                        "<br>" + MainApp.sResources.getString(R.string.canceltemp);
-            } else if (isPercent) {
-                ret += "<br><b>" + MainApp.sResources.getString(R.string.enacted) + "</b>: " + enacted;
-                ret += "<br><b>" + MainApp.sResources.getString(R.string.comment) + "</b>: " + comment;
-                ret += "<br><b>" + MainApp.sResources.getString(R.string.duration) + "</b>: " + duration + " min";
-                ret += "<br><b>" + MainApp.sResources.getString(R.string.percent) + "</b>: " + percent + "%";
-            } else {
-                ret += "<br><b>" + MainApp.sResources.getString(R.string.enacted) + "</b>: " + enacted;
-                ret += "<br><b>" + MainApp.sResources.getString(R.string.comment) + "</b>: " + comment;
-                ret += "<br><b>" + MainApp.sResources.getString(R.string.duration) + "</b>: " + duration + " min";
-                ret += "<br><b>" + MainApp.sResources.getString(R.string.absolute) + "</b>: " + DecimalFormatter.to2Decimal(absolute) + " U/h";
+            if (bolusDelivered > 0) {
+                ret += "<br><b>" + MainApp.gs(R.string.enacted) + "</b>: " + enacted;
+                if (!comment.isEmpty())
+                    ret += "<br><b>" + MainApp.gs(R.string.comment) + "</b>: " + comment;
+                ret += "<br><b>" + MainApp.gs(R.string.smb_shortname) + "</b>: " + bolusDelivered + " " + MainApp.gs(R.string.insulin_unit_shortname);
+            } else if (isTempCancel) {
+                ret += "<br><b>" + MainApp.gs(R.string.enacted) + "</b>: " + enacted;
+                ret += "<br><b>" + MainApp.gs(R.string.comment) + "</b>: " + comment +
+                        "<br>" + MainApp.gs(R.string.canceltemp);
+            } else if (isPercent && percent != -1) {
+                ret += "<br><b>" + MainApp.gs(R.string.enacted) + "</b>: " + enacted;
+                if (!comment.isEmpty())
+                    ret += "<br><b>" + MainApp.gs(R.string.comment) + "</b>: " + comment;
+                ret += "<br><b>" + MainApp.gs(R.string.duration) + "</b>: " + duration + " min";
+                ret += "<br><b>" + MainApp.gs(R.string.percent) + "</b>: " + percent + "%";
+            } else if (absolute != -1) {
+                ret += "<br><b>" + MainApp.gs(R.string.enacted) + "</b>: " + enacted;
+                if (!comment.isEmpty())
+                    ret += "<br><b>" + MainApp.gs(R.string.comment) + "</b>: " + comment;
+                ret += "<br><b>" + MainApp.gs(R.string.duration) + "</b>: " + duration + " min";
+                ret += "<br><b>" + MainApp.gs(R.string.absolute) + "</b>: " + DecimalFormatter.to2Decimal(absolute) + " U/h";
             }
         } else {
-            ret += "<br><b>" + MainApp.sResources.getString(R.string.comment) + "</b>: " + comment;
+            ret += "<br><b>" + MainApp.gs(R.string.comment) + "</b>: " + comment;
         }
-        return Html.fromHtml(ret);
+        return ret;
     }
 
-    public PumpEnactResult() {
-    }
-
-    public JSONObject json() {
+    public JSONObject json(Profile profile) {
         JSONObject result = new JSONObject();
         try {
-            if (isTempCancel) {
+            if (bolusDelivered > 0) {
+                result.put("smb", bolusDelivered);
+            } else if (isTempCancel) {
                 result.put("rate", 0);
                 result.put("duration", 0);
             } else if (isPercent) {
                 // Nightscout is expecting absolute value
-                Double abs = Round.roundTo(MainApp.getConfigBuilder().getProfile().getBasal() * percent / 100, 0.01);
+                Double abs = Round.roundTo(profile.getBasal() * percent / 100, 0.01);
                 result.put("rate", abs);
                 result.put("duration", duration);
             } else {
