@@ -7,18 +7,18 @@ import org.slf4j.LoggerFactory;
 
 import java.util.Date;
 
-import info.nightscout.androidaps.MainApp;
 import info.nightscout.androidaps.R;
 import info.nightscout.androidaps.db.BgReading;
 import info.nightscout.androidaps.db.TempTarget;
 import info.nightscout.androidaps.interfaces.TreatmentsInterface;
-import info.nightscout.androidaps.plugins.ConfigBuilder.ConfigBuilderPlugin;
-import info.nightscout.androidaps.plugins.IobCobCalculator.AutosensData;
-import info.nightscout.androidaps.plugins.IobCobCalculator.IobCobCalculatorPlugin;
-import info.nightscout.androidaps.plugins.Loop.LoopPlugin;
-import info.nightscout.utils.BolusWizard;
-import info.nightscout.utils.DateUtil;
-import info.nightscout.utils.SP;
+import info.nightscout.androidaps.plugins.iob.iobCobCalculator.CobInfo;
+import info.nightscout.androidaps.plugins.iob.iobCobCalculator.GlucoseStatus;
+import info.nightscout.androidaps.plugins.iob.iobCobCalculator.IobCobCalculatorPlugin;
+import info.nightscout.androidaps.plugins.aps.loop.LoopPlugin;
+import info.nightscout.androidaps.plugins.treatments.TreatmentsPlugin;
+import info.nightscout.androidaps.utils.BolusWizard;
+import info.nightscout.androidaps.utils.DateUtil;
+import info.nightscout.androidaps.utils.SP;
 
 /**
  * Created by mike on 25.12.2017.
@@ -80,14 +80,10 @@ public class QuickWizardEntry {
 
         // COB
         double cob = 0d;
-        AutosensData autosensData;
-        if (_synchronized)
-            autosensData = IobCobCalculatorPlugin.getLastAutosensDataSynchronized("QuickWizard COB");
-        else
-            autosensData = IobCobCalculatorPlugin.getLastAutosensData("QuickWizard COB");
-
-        if (autosensData != null && useCOB() == YES) {
-            cob = autosensData.cob;
+        if (useCOB() == YES) {
+            CobInfo cobInfo = IobCobCalculatorPlugin.getPlugin().getCobInfo(_synchronized, "QuickWizard COB");
+            if (cobInfo.displayCob != null)
+                cob = cobInfo.displayCob;
         }
 
         // Temp target
@@ -102,7 +98,7 @@ public class QuickWizardEntry {
         }
 
         // Basal IOB
-        TreatmentsInterface treatments = MainApp.getConfigBuilder();
+        TreatmentsInterface treatments = TreatmentsPlugin.getPlugin();
         treatments.updateTotalIOBTempBasals();
         IobTotal basalIob = treatments.getLastCalculationTempBasals().round();
         boolean basalIOB = false;
@@ -119,8 +115,8 @@ public class QuickWizardEntry {
         if (useSuperBolus() == YES && SP.getBoolean(R.string.key_usesuperbolus, false)) {
             superBolus = true;
         }
-        final LoopPlugin activeloop = ConfigBuilderPlugin.getActiveLoop();
-        if (activeloop != null && activeloop.isEnabled(activeloop.getType()) && activeloop.isSuperBolus())
+        final LoopPlugin loopPlugin = LoopPlugin.getPlugin();
+        if (loopPlugin.isEnabled(loopPlugin.getType()) && loopPlugin.isSuperBolus())
             superBolus = false;
 
         // Trend
