@@ -27,58 +27,68 @@ import java.util.concurrent.ScheduledExecutorService;
 import java.util.concurrent.ScheduledFuture;
 import java.util.concurrent.TimeUnit;
 
-import info.nightscout.androidaps.Config;
 import info.nightscout.androidaps.Constants;
 import info.nightscout.androidaps.MainApp;
-import info.nightscout.androidaps.R;
+import info.nightscout.androidaps.data.OverlappingIntervals;
 import info.nightscout.androidaps.data.Profile;
 import info.nightscout.androidaps.data.ProfileStore;
 import info.nightscout.androidaps.events.EventCareportalEventChange;
 import info.nightscout.androidaps.events.EventExtendedBolusChange;
-import info.nightscout.androidaps.events.EventFoodDatabaseChanged;
 import info.nightscout.androidaps.events.EventNewBG;
-import info.nightscout.androidaps.events.EventProfileSwitchChange;
+import info.nightscout.androidaps.events.EventProfileNeedsUpdate;
 import info.nightscout.androidaps.events.EventRefreshOverview;
 import info.nightscout.androidaps.events.EventReloadProfileSwitchData;
 import info.nightscout.androidaps.events.EventReloadTempBasalData;
 import info.nightscout.androidaps.events.EventReloadTreatmentData;
 import info.nightscout.androidaps.events.EventTempBasalChange;
 import info.nightscout.androidaps.events.EventTempTargetChange;
-import info.nightscout.androidaps.events.EventTreatmentChange;
-import info.nightscout.androidaps.plugins.ConfigBuilder.ConfigBuilderPlugin;
-import info.nightscout.androidaps.plugins.IobCobCalculator.events.EventNewHistoryData;
-import info.nightscout.androidaps.plugins.Overview.events.EventNewNotification;
-import info.nightscout.androidaps.plugins.Overview.notifications.Notification;
-import info.nightscout.androidaps.plugins.PumpDanaR.activities.DanaRNSHistorySync;
-import info.nightscout.androidaps.plugins.PumpVirtual.VirtualPumpPlugin;
-import info.nightscout.utils.DateUtil;
-import info.nightscout.utils.NSUpload;
-import info.nightscout.utils.PercentageSplitter;
+import info.nightscout.androidaps.interfaces.ProfileInterface;
+import info.nightscout.androidaps.logging.L;
+import info.nightscout.androidaps.plugins.configBuilder.ConfigBuilderPlugin;
+import info.nightscout.androidaps.plugins.general.nsclient.NSUpload;
+import info.nightscout.androidaps.plugins.iob.iobCobCalculator.IobCobCalculatorPlugin;
+import info.nightscout.androidaps.plugins.iob.iobCobCalculator.events.EventNewHistoryData;
+import info.nightscout.androidaps.plugins.pump.danaR.activities.DanaRNSHistorySync;
+import info.nightscout.androidaps.plugins.pump.danaR.comm.RecordTypes;
+import info.nightscout.androidaps.plugins.pump.insight.database.InsightBolusID;
+import info.nightscout.androidaps.plugins.pump.insight.database.InsightHistoryOffset;
+import info.nightscout.androidaps.plugins.pump.insight.database.InsightPumpID;
+import info.nightscout.androidaps.plugins.pump.virtual.VirtualPumpPlugin;
+import info.nightscout.androidaps.utils.JsonHelper;
+import info.nightscout.androidaps.utils.PercentageSplitter;
+import info.nightscout.androidaps.utils.ToastUtils;
 
+/**
+ * This Helper contains all resource to provide a central DB management functionality. Only methods handling
+ * data-structure (and not the DB content) should be contained in here (meaning DDL and not SQL).
+ * <p>
+ * This class can safely be called from Services, but should not call Services to avoid circular dependencies.
+ * One major issue with this (right now) are the scheduled events, which are put into the service. Therefor all
+ * direct calls to the corresponding methods (eg. resetDatabases) should be done by a central service.
+ */
 public class DatabaseHelper extends OrmLiteSqliteOpenHelper {
-    private static Logger log = LoggerFactory.getLogger(DatabaseHelper.class);
+    private static Logger log = LoggerFactory.getLogger(L.DATABASE);
 
     public static final String DATABASE_NAME = "AndroidAPSDb";
     public static final String DATABASE_BGREADINGS = "BgReadings";
     public static final String DATABASE_TEMPORARYBASALS = "TemporaryBasals";
     public static final String DATABASE_EXTENDEDBOLUSES = "ExtendedBoluses";
     public static final String DATABASE_TEMPTARGETS = "TempTargets";
-    public static final String DATABASE_TREATMENTS = "Treatments";
     public static final String DATABASE_DANARHISTORY = "DanaRHistory";
     public static final String DATABASE_DBREQUESTS = "DBRequests";
     public static final String DATABASE_CAREPORTALEVENTS = "CareportalEvents";
     public static final String DATABASE_PROFILESWITCHES = "ProfileSwitches";
-    public static final String DATABASE_FOODS = "Foods";
+    public static final String DATABASE_TDDS = "TDDs";
+    public static final String DATABASE_INSIGHT_HISTORY_OFFSETS = "InsightHistoryOffsets";
+    public static final String DATABASE_INSIGHT_BOLUS_IDS = "InsightBolusIDs";
+    public static final String DATABASE_INSIGHT_PUMP_IDS = "InsightPumpIDs";
 
-    private static final int DATABASE_VERSION = 8;
+    private static final int DATABASE_VERSION = 11;
 
-    private static Long earliestDataChange = null;
+    public static Long earliestDataChange = null;
 
     private static final ScheduledExecutorService bgWorker = Executors.newSingleThreadScheduledExecutor();
     private static ScheduledFuture<?> scheduledBgPost = null;
-
-    private static final ScheduledExecutorService treatmentsWorker = Executors.newSingleThreadScheduledExecutor();
-    private static ScheduledFuture<?> scheduledTratmentPost = null;
 
     private static final ScheduledExecutorService tempBasalsWorker = Executors.newSingleThreadScheduledExecutor();
     private static ScheduledFuture<?> scheduledTemBasalsPost = null;
@@ -95,7 +105,8 @@ public class DatabaseHelper extends OrmLiteSqliteOpenHelper {
     private static final ScheduledExecutorService profileSwitchEventWorker = Executors.newSingleThreadScheduledExecutor();
     private static ScheduledFuture<?> scheduledProfileSwitchEventPost = null;
 
-    public FoodHelper foodHelper = new FoodHelper(this);
+    private int oldVersion = 0;
+    private int newVersion = 0;
 
     public DatabaseHelper(Context context) {
         super(context, DATABASE_NAME, null, DATABASE_VERSION);
@@ -106,9 +117,9 @@ public class DatabaseHelper extends OrmLiteSqliteOpenHelper {
     @Override
     public void onCreate(SQLiteDatabase database, ConnectionSource connectionSource) {
         try {
-            log.info("onCreate");
+            if (L.isEnabled(L.DATABASE))
+                log.info("onCreate");
             TableUtils.createTableIfNotExists(connectionSource, TempTarget.class);
-            TableUtils.createTableIfNotExists(connectionSource, Treatment.class);
             TableUtils.createTableIfNotExists(connectionSource, BgReading.class);
             TableUtils.createTableIfNotExists(connectionSource, DanaRHistoryRecord.class);
             TableUtils.createTableIfNotExists(connectionSource, DbRequest.class);
@@ -116,7 +127,14 @@ public class DatabaseHelper extends OrmLiteSqliteOpenHelper {
             TableUtils.createTableIfNotExists(connectionSource, ExtendedBolus.class);
             TableUtils.createTableIfNotExists(connectionSource, CareportalEvent.class);
             TableUtils.createTableIfNotExists(connectionSource, ProfileSwitch.class);
-            TableUtils.createTableIfNotExists(connectionSource, Food.class);
+            TableUtils.createTableIfNotExists(connectionSource, TDD.class);
+            TableUtils.createTableIfNotExists(connectionSource, InsightHistoryOffset.class);
+            TableUtils.createTableIfNotExists(connectionSource, InsightBolusID.class);
+            TableUtils.createTableIfNotExists(connectionSource, InsightPumpID.class);
+            database.execSQL("INSERT INTO sqlite_sequence (name, seq) SELECT \"" + DATABASE_INSIGHT_BOLUS_IDS + "\", " + System.currentTimeMillis() + " " +
+                    "WHERE NOT EXISTS (SELECT 1 FROM sqlite_sequence WHERE name = \"" + DATABASE_INSIGHT_BOLUS_IDS + "\")");
+            database.execSQL("INSERT INTO sqlite_sequence (name, seq) SELECT \"" + DATABASE_INSIGHT_PUMP_IDS + "\", " + System.currentTimeMillis() + " " +
+                    "WHERE NOT EXISTS (SELECT 1 FROM sqlite_sequence WHERE name = \"" + DATABASE_INSIGHT_PUMP_IDS + "\")");
         } catch (SQLException e) {
             log.error("Can't create database", e);
             throw new RuntimeException(e);
@@ -126,14 +144,12 @@ public class DatabaseHelper extends OrmLiteSqliteOpenHelper {
     @Override
     public void onUpgrade(SQLiteDatabase database, ConnectionSource connectionSource, int oldVersion, int newVersion) {
         try {
-            if (oldVersion == 7 && newVersion == 8) {
-                log.debug("Upgrading database from v7 to v8");
-                TableUtils.dropTable(connectionSource, Treatment.class, true);
-                TableUtils.createTableIfNotExists(connectionSource, Treatment.class);
-            } else {
+            this.oldVersion = oldVersion;
+            this.newVersion = newVersion;
+
+            if (oldVersion < 7) {
                 log.info(DatabaseHelper.class.getName(), "onUpgrade");
                 TableUtils.dropTable(connectionSource, TempTarget.class, true);
-                TableUtils.dropTable(connectionSource, Treatment.class, true);
                 TableUtils.dropTable(connectionSource, BgReading.class, true);
                 TableUtils.dropTable(connectionSource, DanaRHistoryRecord.class, true);
                 TableUtils.dropTable(connectionSource, DbRequest.class, true);
@@ -141,13 +157,37 @@ public class DatabaseHelper extends OrmLiteSqliteOpenHelper {
                 TableUtils.dropTable(connectionSource, ExtendedBolus.class, true);
                 TableUtils.dropTable(connectionSource, CareportalEvent.class, true);
                 TableUtils.dropTable(connectionSource, ProfileSwitch.class, true);
-                TableUtils.dropTable(connectionSource, Food.class, true);
                 onCreate(database, connectionSource);
+            } else if (oldVersion < 10) {
+                TableUtils.createTableIfNotExists(connectionSource, InsightHistoryOffset.class);
+                TableUtils.createTableIfNotExists(connectionSource, InsightBolusID.class);
+                TableUtils.createTableIfNotExists(connectionSource, InsightPumpID.class);
+                database.execSQL("INSERT INTO sqlite_sequence (name, seq) SELECT \"" + DATABASE_INSIGHT_BOLUS_IDS + "\", " + System.currentTimeMillis() + " " +
+                        "WHERE NOT EXISTS (SELECT 1 FROM sqlite_sequence WHERE name = \"" + DATABASE_INSIGHT_BOLUS_IDS + "\")");
+                database.execSQL("INSERT INTO sqlite_sequence (name, seq) SELECT \"" + DATABASE_INSIGHT_PUMP_IDS + "\", " + System.currentTimeMillis() + " " +
+                        "WHERE NOT EXISTS (SELECT 1 FROM sqlite_sequence WHERE name = \"" + DATABASE_INSIGHT_PUMP_IDS + "\")");
+            } else if (oldVersion < 11) {
+                database.execSQL("UPDATE sqlite_sequence SET seq = " + System.currentTimeMillis() + " WHERE name = \"" + DATABASE_INSIGHT_BOLUS_IDS + "\"");
+                database.execSQL("UPDATE sqlite_sequence SET seq = " + System.currentTimeMillis() + " WHERE name = \"" + DATABASE_INSIGHT_PUMP_IDS + "\"");
             }
         } catch (SQLException e) {
             log.error("Can't drop databases", e);
             throw new RuntimeException(e);
         }
+    }
+
+    @Override
+    public void onDowngrade(SQLiteDatabase db, int oldVersion, int newVersion) {
+        log.info("Do nothing for downgrading...");
+        log.debug("oldVersion: {}, newVersion: {}", oldVersion, newVersion);
+    }
+
+    public int getOldVersion() {
+        return oldVersion;
+    }
+
+    public int getNewVersion() {
+        return newVersion;
     }
 
     /**
@@ -158,40 +198,6 @@ public class DatabaseHelper extends OrmLiteSqliteOpenHelper {
         super.close();
     }
 
-    public void cleanUpDatabases() {
-        // TODO: call it somewhere
-        log.debug("Before BgReadings size: " + DatabaseUtils.queryNumEntries(getReadableDatabase(), DATABASE_BGREADINGS));
-        getWritableDatabase().delete(DATABASE_BGREADINGS, "date" + " < '" + (System.currentTimeMillis() - Constants.hoursToKeepInDatabase * 60 * 60 * 1000L) + "'", null);
-        log.debug("After BgReadings size: " + DatabaseUtils.queryNumEntries(getReadableDatabase(), DATABASE_BGREADINGS));
-
-        log.debug("Before TempTargets size: " + DatabaseUtils.queryNumEntries(getReadableDatabase(), DATABASE_TEMPTARGETS));
-        getWritableDatabase().delete(DATABASE_TEMPTARGETS, "date" + " < '" + (System.currentTimeMillis() - Constants.hoursToKeepInDatabase * 60 * 60 * 1000L) + "'", null);
-        log.debug("After TempTargets size: " + DatabaseUtils.queryNumEntries(getReadableDatabase(), DATABASE_TEMPTARGETS));
-
-        log.debug("Before Treatments size: " + DatabaseUtils.queryNumEntries(getReadableDatabase(), DATABASE_TREATMENTS));
-        getWritableDatabase().delete(DATABASE_TREATMENTS, "date" + " < '" + (System.currentTimeMillis() - Constants.hoursToKeepInDatabase * 60 * 60 * 1000L) + "'", null);
-        log.debug("After Treatments size: " + DatabaseUtils.queryNumEntries(getReadableDatabase(), DATABASE_TREATMENTS));
-
-        log.debug("Before History size: " + DatabaseUtils.queryNumEntries(getReadableDatabase(), DATABASE_DANARHISTORY));
-        getWritableDatabase().delete(DATABASE_DANARHISTORY, "recordDate" + " < '" + (System.currentTimeMillis() - Constants.daysToKeepHistoryInDatabase * 24 * 60 * 60 * 1000L) + "'", null);
-        log.debug("After History size: " + DatabaseUtils.queryNumEntries(getReadableDatabase(), DATABASE_DANARHISTORY));
-
-        log.debug("Before TemporaryBasals size: " + DatabaseUtils.queryNumEntries(getReadableDatabase(), DATABASE_TEMPORARYBASALS));
-        getWritableDatabase().delete(DATABASE_TEMPORARYBASALS, "recordDate" + " < '" + (System.currentTimeMillis() - Constants.daysToKeepHistoryInDatabase * 24 * 60 * 60 * 1000L) + "'", null);
-        log.debug("After TemporaryBasals size: " + DatabaseUtils.queryNumEntries(getReadableDatabase(), DATABASE_TEMPORARYBASALS));
-
-        log.debug("Before ExtendedBoluses size: " + DatabaseUtils.queryNumEntries(getReadableDatabase(), DATABASE_EXTENDEDBOLUSES));
-        getWritableDatabase().delete(DATABASE_EXTENDEDBOLUSES, "recordDate" + " < '" + (System.currentTimeMillis() - Constants.daysToKeepHistoryInDatabase * 24 * 60 * 60 * 1000L) + "'", null);
-        log.debug("After ExtendedBoluses size: " + DatabaseUtils.queryNumEntries(getReadableDatabase(), DATABASE_EXTENDEDBOLUSES));
-
-        log.debug("Before CareportalEvent size: " + DatabaseUtils.queryNumEntries(getReadableDatabase(), DATABASE_CAREPORTALEVENTS));
-        getWritableDatabase().delete(DATABASE_CAREPORTALEVENTS, "recordDate" + " < '" + (System.currentTimeMillis() - Constants.daysToKeepHistoryInDatabase * 24 * 60 * 60 * 1000L) + "'", null);
-        log.debug("After CareportalEvent size: " + DatabaseUtils.queryNumEntries(getReadableDatabase(), DATABASE_CAREPORTALEVENTS));
-
-        log.debug("Before ProfileSwitch size: " + DatabaseUtils.queryNumEntries(getReadableDatabase(), DATABASE_PROFILESWITCHES));
-        getWritableDatabase().delete(DATABASE_PROFILESWITCHES, "recordDate" + " < '" + (System.currentTimeMillis() - Constants.daysToKeepHistoryInDatabase * 24 * 60 * 60 * 1000L) + "'", null);
-        log.debug("After ProfileSwitch size: " + DatabaseUtils.queryNumEntries(getReadableDatabase(), DATABASE_PROFILESWITCHES));
-    }
 
     public long size(String database) {
         return DatabaseUtils.queryNumEntries(getReadableDatabase(), database);
@@ -202,7 +208,6 @@ public class DatabaseHelper extends OrmLiteSqliteOpenHelper {
     public void resetDatabases() {
         try {
             TableUtils.dropTable(connectionSource, TempTarget.class, true);
-            TableUtils.dropTable(connectionSource, Treatment.class, true);
             TableUtils.dropTable(connectionSource, BgReading.class, true);
             TableUtils.dropTable(connectionSource, DanaRHistoryRecord.class, true);
             TableUtils.dropTable(connectionSource, DbRequest.class, true);
@@ -210,8 +215,8 @@ public class DatabaseHelper extends OrmLiteSqliteOpenHelper {
             TableUtils.dropTable(connectionSource, ExtendedBolus.class, true);
             TableUtils.dropTable(connectionSource, CareportalEvent.class, true);
             TableUtils.dropTable(connectionSource, ProfileSwitch.class, true);
+            TableUtils.dropTable(connectionSource, TDD.class, true);
             TableUtils.createTableIfNotExists(connectionSource, TempTarget.class);
-            TableUtils.createTableIfNotExists(connectionSource, Treatment.class);
             TableUtils.createTableIfNotExists(connectionSource, BgReading.class);
             TableUtils.createTableIfNotExists(connectionSource, DanaRHistoryRecord.class);
             TableUtils.createTableIfNotExists(connectionSource, DbRequest.class);
@@ -219,20 +224,18 @@ public class DatabaseHelper extends OrmLiteSqliteOpenHelper {
             TableUtils.createTableIfNotExists(connectionSource, ExtendedBolus.class);
             TableUtils.createTableIfNotExists(connectionSource, CareportalEvent.class);
             TableUtils.createTableIfNotExists(connectionSource, ProfileSwitch.class);
-            foodHelper.resetFood();
+            TableUtils.createTableIfNotExists(connectionSource, TDD.class);
             updateEarliestDataChange(0);
         } catch (SQLException e) {
             log.error("Unhandled exception", e);
         }
-        VirtualPumpPlugin.setFakingStatus(true);
-        scheduleBgChange(); // trigger refresh
+        VirtualPumpPlugin.getPlugin().setFakingStatus(true);
+        scheduleBgChange(null); // trigger refresh
         scheduleTemporaryBasalChange();
-        scheduleTreatmentChange();
         scheduleExtendedBolusChange();
         scheduleTemporaryTargetChange();
         scheduleCareportalEventChange();
         scheduleProfileSwitchChange();
-        foodHelper.scheduleFoodChange();
         new java.util.Timer().schedule(
                 new java.util.TimerTask() {
                     @Override
@@ -242,17 +245,6 @@ public class DatabaseHelper extends OrmLiteSqliteOpenHelper {
                 },
                 3000
         );
-    }
-
-    public void resetTreatments() {
-        try {
-            TableUtils.dropTable(connectionSource, Treatment.class, true);
-            TableUtils.createTableIfNotExists(connectionSource, Treatment.class);
-            updateEarliestDataChange(0);
-        } catch (SQLException e) {
-            log.error("Unhandled exception", e);
-        }
-        scheduleTreatmentChange();
     }
 
     public void resetTempTargets() {
@@ -273,7 +265,7 @@ public class DatabaseHelper extends OrmLiteSqliteOpenHelper {
         } catch (SQLException e) {
             log.error("Unhandled exception", e);
         }
-        VirtualPumpPlugin.setFakingStatus(false);
+        VirtualPumpPlugin.getPlugin().setFakingStatus(false);
         scheduleTemporaryBasalChange();
     }
 
@@ -308,14 +300,19 @@ public class DatabaseHelper extends OrmLiteSqliteOpenHelper {
         scheduleProfileSwitchChange();
     }
 
+    public void resetTDDs() {
+        try {
+            TableUtils.dropTable(connectionSource, TDD.class, true);
+            TableUtils.createTableIfNotExists(connectionSource, TDD.class);
+        } catch (SQLException e) {
+            log.error("Unhandled exception", e);
+        }
+    }
+
     // ------------------ getDao -------------------------------------------
 
     private Dao<TempTarget, Long> getDaoTempTargets() throws SQLException {
         return getDao(TempTarget.class);
-    }
-
-    private Dao<Treatment, Long> getDaoTreatments() throws SQLException {
-        return getDao(Treatment.class);
     }
 
     private Dao<BgReading, Long> getDaoBgReadings() throws SQLException {
@@ -324,6 +321,10 @@ public class DatabaseHelper extends OrmLiteSqliteOpenHelper {
 
     private Dao<DanaRHistoryRecord, String> getDaoDanaRHistory() throws SQLException {
         return getDao(DanaRHistoryRecord.class);
+    }
+
+    private Dao<TDD, String> getDaoTDD() throws SQLException {
+        return getDao(TDD.class);
     }
 
     private Dao<DbRequest, String> getDaoDbRequest() throws SQLException {
@@ -346,8 +347,24 @@ public class DatabaseHelper extends OrmLiteSqliteOpenHelper {
         return getDao(ProfileSwitch.class);
     }
 
-    public long roundDateToSec(long date) {
-        return date - date % 1000;
+    private Dao<InsightPumpID, Long> getDaoInsightPumpID() throws SQLException {
+        return getDao(InsightPumpID.class);
+    }
+
+    private Dao<InsightBolusID, Long> getDaoInsightBolusID() throws SQLException {
+        return getDao(InsightBolusID.class);
+    }
+
+    private Dao<InsightHistoryOffset, String> getDaoInsightHistoryOffset() throws SQLException {
+        return getDao(InsightHistoryOffset.class);
+    }
+
+    public static long roundDateToSec(long date) {
+        long rounded = date - date % 1000;
+        if (rounded != date)
+            if (L.isEnabled(L.DATABASE))
+                log.debug("Rounding " + date + " to " + rounded);
+        return rounded;
     }
     // -------------------  BgReading handling -----------------------
 
@@ -357,16 +374,19 @@ public class DatabaseHelper extends OrmLiteSqliteOpenHelper {
             BgReading old = getDaoBgReadings().queryForId(bgReading.date);
             if (old == null) {
                 getDaoBgReadings().create(bgReading);
-                log.debug("BG: New record from: " + from + " " + bgReading.toString());
-                scheduleBgChange();
+                if (L.isEnabled(L.DATABASE))
+                    log.debug("BG: New record from: " + from + " " + bgReading.toString());
+                scheduleBgChange(bgReading);
                 return true;
             }
             if (!old.isEqual(bgReading)) {
-                log.debug("BG: Similiar found: " + old.toString());
+                if (L.isEnabled(L.DATABASE))
+                    log.debug("BG: Similiar found: " + old.toString());
                 old.copyFrom(bgReading);
                 getDaoBgReadings().update(old);
-                log.debug("BG: Updating record from: " + from + " New data: " + old.toString());
-                scheduleBgChange();
+                if (L.isEnabled(L.DATABASE))
+                    log.debug("BG: Updating record from: " + from + " New data: " + old.toString());
+                scheduleBgChange(bgReading);
                 return false;
             }
         } catch (SQLException e) {
@@ -380,15 +400,16 @@ public class DatabaseHelper extends OrmLiteSqliteOpenHelper {
         try {
             getDaoBgReadings().update(bgReading);
         } catch (SQLException e) {
-            e.printStackTrace();
+            log.error("Unhandled exception", e);
         }
     }
 
-    private static void scheduleBgChange() {
+    private static void scheduleBgChange(@Nullable final BgReading bgReading) {
         class PostRunnable implements Runnable {
             public void run() {
-                log.debug("Firing EventNewBg");
-                MainApp.bus().post(new EventNewBG());
+                if (L.isEnabled(L.DATABASE))
+                    log.debug("Firing EventNewBg");
+                MainApp.bus().post(new EventNewBG(bgReading));
                 scheduledBgPost = null;
             }
         }
@@ -403,34 +424,25 @@ public class DatabaseHelper extends OrmLiteSqliteOpenHelper {
     }
 
     /*
-         * Return last BgReading from database or null if db is empty
-         */
+     * Return last BgReading from database or null if db is empty
+     */
     @Nullable
     public static BgReading lastBg() {
-        List<BgReading> bgList = null;
+        List<BgReading> bgList = IobCobCalculatorPlugin.getPlugin().getBgReadings();
 
-        try {
-            Dao<BgReading, Long> daoBgReadings = MainApp.getDbHelper().getDaoBgReadings();
-            QueryBuilder<BgReading, Long> queryBuilder = daoBgReadings.queryBuilder();
-            queryBuilder.orderBy("date", false);
-            queryBuilder.limit(1L);
-            queryBuilder.where().gt("value", 38).and().eq("isValid", true);
-            PreparedQuery<BgReading> preparedQuery = queryBuilder.prepare();
-            bgList = daoBgReadings.query(preparedQuery);
-
-        } catch (SQLException e) {
-            log.debug(e.getMessage(), e);
-        }
-        if (bgList != null && bgList.size() > 0)
-            return bgList.get(0);
-        else
+        if (bgList == null)
             return null;
+
+        for (int i = 0; i < bgList.size(); i++)
+            if (bgList.get(i).value > 39)
+                return bgList.get(i);
+        return null;
     }
 
     /*
-         * Return bg reading if not old ( <9 min )
-         * or null if older
-         */
+     * Return bg reading if not old ( <9 min )
+     * or null if older
+     */
     @Nullable
     public static BgReading actualBg() {
         BgReading lastBg = lastBg();
@@ -452,14 +464,31 @@ public class DatabaseHelper extends OrmLiteSqliteOpenHelper {
             QueryBuilder<BgReading, Long> queryBuilder = daoBgreadings.queryBuilder();
             queryBuilder.orderBy("date", ascending);
             Where where = queryBuilder.where();
-            where.ge("date", mills).and().gt("value", 38).and().eq("isValid", true);
+            where.ge("date", mills).and().ge("value", 39).and().eq("isValid", true);
             PreparedQuery<BgReading> preparedQuery = queryBuilder.prepare();
             bgReadings = daoBgreadings.query(preparedQuery);
             return bgReadings;
         } catch (SQLException e) {
             log.error("Unhandled exception", e);
         }
-        return new ArrayList<BgReading>();
+        return new ArrayList<>();
+    }
+
+    public List<BgReading> getBgreadingsDataFromTime(long start, long end, boolean ascending) {
+        try {
+            Dao<BgReading, Long> daoBgreadings = getDaoBgReadings();
+            List<BgReading> bgReadings;
+            QueryBuilder<BgReading, Long> queryBuilder = daoBgreadings.queryBuilder();
+            queryBuilder.orderBy("date", ascending);
+            Where where = queryBuilder.where();
+            where.between("date", start, end).and().ge("value", 39).and().eq("isValid", true);
+            PreparedQuery<BgReading> preparedQuery = queryBuilder.prepare();
+            bgReadings = daoBgreadings.query(preparedQuery);
+            return bgReadings;
+        } catch (SQLException e) {
+            log.error("Unhandled exception", e);
+        }
+        return new ArrayList<>();
     }
 
     public List<BgReading> getAllBgreadingsDataFromTime(long mills, boolean ascending) {
@@ -478,6 +507,33 @@ public class DatabaseHelper extends OrmLiteSqliteOpenHelper {
         }
         return new ArrayList<BgReading>();
     }
+
+    // -------------------  TDD handling -----------------------
+    public void createOrUpdateTDD(TDD tdd) {
+        try {
+            Dao<TDD, String> dao = getDaoTDD();
+            dao.createOrUpdate(tdd);
+        } catch (SQLException e) {
+            ToastUtils.showToastInUiThread(MainApp.instance(), "createOrUpdate-Exception");
+            log.error("Unhandled exception", e);
+        }
+    }
+
+    public List<TDD> getTDDs() {
+        List<TDD> tddList;
+        try {
+            QueryBuilder<TDD, String> queryBuilder = getDaoTDD().queryBuilder();
+            queryBuilder.orderBy("date", false);
+            queryBuilder.limit(10L);
+            PreparedQuery<TDD> preparedQuery = queryBuilder.prepare();
+            tddList = getDaoTDD().query(preparedQuery);
+        } catch (SQLException e) {
+            log.error("Unhandled exception", e);
+            tddList = new ArrayList<>();
+        }
+        return tddList;
+    }
+
 
     // ------------- DbRequests handling -------------------
 
@@ -515,7 +571,6 @@ public class DatabaseHelper extends OrmLiteSqliteOpenHelper {
             queryBuilder.limit(10L);
             PreparedQuery<DbRequest> preparedQuery = queryBuilder.prepare();
             List<DbRequest> dbList = getDaoDbRequest().query(preparedQuery);
-            log.error("deleteDbRequestbyMongoId query size: " + dbList.size());
             for (DbRequest r : dbList) {
                 delete(r);
             }
@@ -543,219 +598,13 @@ public class DatabaseHelper extends OrmLiteSqliteOpenHelper {
 
     //  -------------------- TREATMENT HANDLING -------------------
 
-    // return true if new record is created
-    public boolean createOrUpdate(Treatment treatment) {
-        try {
-            Treatment old;
-            treatment.date = roundDateToSec(treatment.date);
-
-            if (treatment.source == Source.PUMP) {
-                // check for changed from pump change in NS
-                QueryBuilder<Treatment, Long> queryBuilder = getDaoTreatments().queryBuilder();
-                Where where = queryBuilder.where();
-                where.eq("pumpId", treatment.pumpId);
-                PreparedQuery<Treatment> preparedQuery = queryBuilder.prepare();
-                List<Treatment> trList = getDaoTreatments().query(preparedQuery);
-                if (trList.size() > 0) {
-                    // do nothing, pump history record cannot be changed
-                    return false;
-                }
-                getDaoTreatments().create(treatment);
-                log.debug("TREATMENT: New record from: " + Source.getString(treatment.source) + " " + treatment.toString());
-                updateEarliestDataChange(treatment.date);
-                scheduleTreatmentChange();
-                return true;
-            }
-            if (treatment.source == Source.NIGHTSCOUT) {
-                old = getDaoTreatments().queryForId(treatment.date);
-                if (old != null) {
-                    if (!old.isEqual(treatment)) {
-                        boolean historyChange = old.isDataChanging(treatment);
-                        long oldDate = old.date;
-                        getDaoTreatments().delete(old); // need to delete/create because date may change too
-                        old.copyFrom(treatment);
-                        getDaoTreatments().create(old);
-                        log.debug("TREATMENT: Updating record by date from: " + Source.getString(treatment.source) + " " + old.toString());
-                        if (historyChange) {
-                            updateEarliestDataChange(oldDate);
-                            updateEarliestDataChange(old.date);
-                        }
-                        scheduleTreatmentChange();
-                        return true;
-                    }
-                    return false;
-                }
-                // find by NS _id
-                if (treatment._id != null) {
-                    QueryBuilder<Treatment, Long> queryBuilder = getDaoTreatments().queryBuilder();
-                    Where where = queryBuilder.where();
-                    where.eq("_id", treatment._id);
-                    PreparedQuery<Treatment> preparedQuery = queryBuilder.prepare();
-                    List<Treatment> trList = getDaoTreatments().query(preparedQuery);
-                    if (trList.size() > 0) {
-                        old = trList.get(0);
-                        if (!old.isEqual(treatment)) {
-                            boolean historyChange = old.isDataChanging(treatment);
-                            long oldDate = old.date;
-                            getDaoTreatments().delete(old); // need to delete/create because date may change too
-                            old.copyFrom(treatment);
-                            getDaoTreatments().create(old);
-                            log.debug("TREATMENT: Updating record by _id from: " + Source.getString(treatment.source) + " " + old.toString());
-                            if (historyChange) {
-                                updateEarliestDataChange(oldDate);
-                                updateEarliestDataChange(old.date);
-                            }
-                            scheduleTreatmentChange();
-                            return true;
-                        }
-                    }
-                }
-                getDaoTreatments().create(treatment);
-                log.debug("TREATMENT: New record from: " + Source.getString(treatment.source) + " " + treatment.toString());
-                updateEarliestDataChange(treatment.date);
-                scheduleTreatmentChange();
-                return true;
-            }
-            if (treatment.source == Source.USER) {
-                getDaoTreatments().create(treatment);
-                log.debug("TREATMENT: New record from: " + Source.getString(treatment.source) + " " + treatment.toString());
-                updateEarliestDataChange(treatment.date);
-                scheduleTreatmentChange();
-                return true;
-            }
-        } catch (SQLException e) {
-            log.error("Unhandled exception", e);
-        }
-        return false;
-    }
-
-    public void delete(Treatment treatment) {
-        try {
-            getDaoTreatments().delete(treatment);
-            updateEarliestDataChange(treatment.date);
-        } catch (SQLException e) {
-            log.error("Unhandled exception", e);
-        }
-        scheduleTreatmentChange();
-    }
-
-    public void update(Treatment treatment) {
-        try {
-            getDaoTreatments().update(treatment);
-            updateEarliestDataChange(treatment.date);
-        } catch (SQLException e) {
-            log.error("Unhandled exception", e);
-        }
-        scheduleTreatmentChange();
-    }
-
-    public void deleteTreatmentById(String _id) {
-        Treatment stored = findTreatmentById(_id);
-        if (stored != null) {
-            log.debug("TREATMENT: Removing Treatment record from database: " + stored.toString());
-            delete(stored);
-            updateEarliestDataChange(stored.date);
-            scheduleTreatmentChange();
-        }
-    }
-
-    @Nullable
-    public Treatment findTreatmentById(String _id) {
-        try {
-            Dao<Treatment, Long> daoTreatments = getDaoTreatments();
-            QueryBuilder<Treatment, Long> queryBuilder = daoTreatments.queryBuilder();
-            Where where = queryBuilder.where();
-            where.eq("_id", _id);
-            queryBuilder.limit(10L);
-            PreparedQuery<Treatment> preparedQuery = queryBuilder.prepare();
-            List<Treatment> trList = daoTreatments.query(preparedQuery);
-            if (trList.size() != 1) {
-                //log.debug("Treatment findTreatmentById query size: " + trList.size());
-                return null;
-            } else {
-                //log.debug("Treatment findTreatmentById found: " + trList.get(0).log());
-                return trList.get(0);
-            }
-        } catch (SQLException e) {
-            log.error("Unhandled exception", e);
-        }
-        return null;
-    }
-
-    private void updateEarliestDataChange(long newDate) {
+    public static void updateEarliestDataChange(long newDate) {
         if (earliestDataChange == null) {
             earliestDataChange = newDate;
             return;
         }
         if (newDate < earliestDataChange) {
             earliestDataChange = newDate;
-        }
-    }
-
-    private static void scheduleTreatmentChange() {
-        class PostRunnable implements Runnable {
-            public void run() {
-                log.debug("Firing EventTreatmentChange");
-                MainApp.bus().post(new EventReloadTreatmentData(new EventTreatmentChange()));
-                if (earliestDataChange != null)
-                    MainApp.bus().post(new EventNewHistoryData(earliestDataChange));
-                earliestDataChange = null;
-                scheduledTratmentPost = null;
-            }
-        }
-        // prepare task for execution in 1 sec
-        // cancel waiting task to prevent sending multiple posts
-        if (scheduledTratmentPost != null)
-            scheduledTratmentPost.cancel(false);
-        Runnable task = new PostRunnable();
-        final int sec = 1;
-        scheduledTratmentPost = treatmentsWorker.schedule(task, sec, TimeUnit.SECONDS);
-
-    }
-
-    public List<Treatment> getTreatmentDataFromTime(long mills, boolean ascending) {
-        try {
-            Dao<Treatment, Long> daoTreatments = getDaoTreatments();
-            List<Treatment> treatments;
-            QueryBuilder<Treatment, Long> queryBuilder = daoTreatments.queryBuilder();
-            queryBuilder.orderBy("date", ascending);
-            Where where = queryBuilder.where();
-            where.ge("date", mills);
-            PreparedQuery<Treatment> preparedQuery = queryBuilder.prepare();
-            treatments = daoTreatments.query(preparedQuery);
-            return treatments;
-        } catch (SQLException e) {
-            log.error("Unhandled exception", e);
-        }
-        return new ArrayList<Treatment>();
-    }
-
-    public void createTreatmentFromJsonIfNotExists(JSONObject trJson) {
-        try {
-            Treatment treatment = new Treatment();
-            treatment.source = Source.NIGHTSCOUT;
-            treatment.date = roundDateToSec(trJson.getLong("mills"));
-            treatment.carbs = trJson.has("carbs") ? trJson.getDouble("carbs") : 0;
-            treatment.insulin = trJson.has("insulin") ? trJson.getDouble("insulin") : 0d;
-            treatment.pumpId = trJson.has("pumpId") ? trJson.getLong("pumpId") : 0;
-            treatment._id = trJson.getString("_id");
-            if (trJson.has("isSMB"))
-                treatment.isSMB = trJson.getBoolean("isSMB");
-            if (trJson.has("eventType")) {
-                treatment.mealBolus = !trJson.get("eventType").equals("Correction Bolus");
-                double carbs = treatment.carbs;
-                if (trJson.has("boluscalc")) {
-                    JSONObject boluscalc = trJson.getJSONObject("boluscalc");
-                    if (boluscalc.has("carbs")) {
-                        carbs = Math.max(boluscalc.getDouble("carbs"), carbs);
-                    }
-                }
-                if (carbs <= 0)
-                    treatment.mealBolus = false;
-            }
-            createOrUpdate(treatment);
-        } catch (JSONException e) {
-            log.error("Unhandled exception", e);
         }
     }
 
@@ -790,7 +639,8 @@ public class DatabaseHelper extends OrmLiteSqliteOpenHelper {
                         getDaoTempTargets().delete(old); // need to delete/create because date may change too
                         old.copyFrom(tempTarget);
                         getDaoTempTargets().create(old);
-                        log.debug("TEMPTARGET: Updating record by date from: " + Source.getString(tempTarget.source) + " " + old.toString());
+                        if (L.isEnabled(L.DATABASE))
+                            log.debug("TEMPTARGET: Updating record by date from: " + Source.getString(tempTarget.source) + " " + old.toString());
                         scheduleTemporaryTargetChange();
                         return true;
                     }
@@ -809,20 +659,23 @@ public class DatabaseHelper extends OrmLiteSqliteOpenHelper {
                             getDaoTempTargets().delete(old); // need to delete/create because date may change too
                             old.copyFrom(tempTarget);
                             getDaoTempTargets().create(old);
-                            log.debug("TEMPTARGET: Updating record by _id from: " + Source.getString(tempTarget.source) + " " + old.toString());
+                            if (L.isEnabled(L.DATABASE))
+                                log.debug("TEMPTARGET: Updating record by _id from: " + Source.getString(tempTarget.source) + " " + old.toString());
                             scheduleTemporaryTargetChange();
                             return true;
                         }
                     }
                 }
                 getDaoTempTargets().create(tempTarget);
-                log.debug("TEMPTARGET: New record from: " + Source.getString(tempTarget.source) + " " + tempTarget.toString());
+                if (L.isEnabled(L.DATABASE))
+                    log.debug("TEMPTARGET: New record from: " + Source.getString(tempTarget.source) + " " + tempTarget.toString());
                 scheduleTemporaryTargetChange();
                 return true;
             }
             if (tempTarget.source == Source.USER) {
                 getDaoTempTargets().create(tempTarget);
-                log.debug("TEMPTARGET: New record from: " + Source.getString(tempTarget.source) + " " + tempTarget.toString());
+                if (L.isEnabled(L.DATABASE))
+                    log.debug("TEMPTARGET: New record from: " + Source.getString(tempTarget.source) + " " + tempTarget.toString());
                 scheduleTemporaryTargetChange();
                 return true;
             }
@@ -844,7 +697,8 @@ public class DatabaseHelper extends OrmLiteSqliteOpenHelper {
     private static void scheduleTemporaryTargetChange() {
         class PostRunnable implements Runnable {
             public void run() {
-                log.debug("Firing EventTempTargetChange");
+                if (L.isEnabled(L.DATABASE))
+                    log.debug("Firing EventTempTargetChange");
                 MainApp.bus().post(new EventTempTargetChange());
                 scheduledTemTargetPost = null;
             }
@@ -876,18 +730,18 @@ public class DatabaseHelper extends OrmLiteSqliteOpenHelper {
 
     public void createTemptargetFromJsonIfNotExists(JSONObject trJson) {
         try {
-            String units = MainApp.getConfigBuilder().getProfileUnits();
-            TempTarget tempTarget = new TempTarget();
-            tempTarget.date = trJson.getLong("mills");
-            tempTarget.durationInMinutes = trJson.getInt("duration");
-            tempTarget.low = Profile.toMgdl(trJson.getDouble("targetBottom"), units);
-            tempTarget.high = Profile.toMgdl(trJson.getDouble("targetTop"), units);
-            tempTarget.reason = trJson.getString("reason");
-            tempTarget._id = trJson.getString("_id");
-            tempTarget.source = Source.NIGHTSCOUT;
+            String units = JsonHelper.safeGetString(trJson, "units", Constants.MGDL);
+            TempTarget tempTarget = new TempTarget()
+                    .date(trJson.getLong("mills"))
+                    .duration(JsonHelper.safeGetInt(trJson, "duration"))
+                    .low(Profile.toMgdl(trJson.getDouble("targetBottom"), units))
+                    .high(Profile.toMgdl(trJson.getDouble("targetTop"), units))
+                    .reason(JsonHelper.safeGetString(trJson, "reason", ""))
+                    ._id(trJson.getString("_id"))
+                    .source(Source.NIGHTSCOUT);
             createOrUpdate(tempTarget);
         } catch (JSONException e) {
-            log.error("Unhandled exception", e);
+            log.error("Unhandled exception: " + trJson.toString(), e);
         }
     }
 
@@ -924,6 +778,12 @@ public class DatabaseHelper extends OrmLiteSqliteOpenHelper {
     public void createOrUpdate(DanaRHistoryRecord record) {
         try {
             getDaoDanaRHistory().createOrUpdate(record);
+
+            //If it is a TDD, store it for stats also.
+            if (record.recordCode == RecordTypes.RECORD_TYPE_DAILY) {
+                createOrUpdateTDD(new TDD(record.recordDate, record.recordDailyBolus, record.recordDailyBasal, 0));
+            }
+
         } catch (SQLException e) {
             log.error("Unhandled exception", e);
         }
@@ -958,7 +818,7 @@ public class DatabaseHelper extends OrmLiteSqliteOpenHelper {
             } else if (list.size() == 1) {
                 DanaRHistoryRecord record = list.get(0);
                 if (record._id == null || !record._id.equals(trJson.getString("_id"))) {
-                    if (Config.logIncommingData)
+                    if (L.isEnabled(L.DATABASE))
                         log.debug("Updating _id in DanaR history database: " + trJson.getString("_id"));
                     record._id = trJson.getString("_id");
                     getDaoDanaRHistory().update(record);
@@ -967,7 +827,7 @@ public class DatabaseHelper extends OrmLiteSqliteOpenHelper {
                 }
             }
         } catch (SQLException | JSONException e) {
-            log.error("Unhandled exception", e);
+            log.error("Unhandled exception: " + trJson.toString(), e);
         }
     }
 
@@ -988,11 +848,13 @@ public class DatabaseHelper extends OrmLiteSqliteOpenHelper {
                 List<TemporaryBasal> trList = getDaoTemporaryBasal().query(preparedQuery);
                 if (trList.size() > 0) {
                     // do nothing, pump history record cannot be changed
-                    log.debug("TEMPBASAL: Already exists from: " + Source.getString(tempBasal.source) + " " + tempBasal.toString());
+                    if (L.isEnabled(L.DATABASE))
+                        log.debug("TEMPBASAL: Already exists from: " + Source.getString(tempBasal.source) + " " + tempBasal.toString());
                     return false;
                 }
                 getDaoTemporaryBasal().create(tempBasal);
-                log.debug("TEMPBASAL: New record from: " + Source.getString(tempBasal.source) + " " + tempBasal.toString());
+                if (L.isEnabled(L.DATABASE))
+                    log.debug("TEMPBASAL: New record from: " + Source.getString(tempBasal.source) + " " + tempBasal.toString());
                 updateEarliestDataChange(tempBasal.date);
                 scheduleTemporaryBasalChange();
                 return true;
@@ -1009,7 +871,8 @@ public class DatabaseHelper extends OrmLiteSqliteOpenHelper {
                         getDaoTemporaryBasal().delete(old); // need to delete/create because date may change too
                         old.copyFrom(tempBasal);
                         getDaoTemporaryBasal().create(old);
-                        log.debug("TEMPBASAL: Updating record by date from: " + Source.getString(tempBasal.source) + " " + old.toString());
+                        if (L.isEnabled(L.DATABASE))
+                            log.debug("TEMPBASAL: Updating record by date from: " + Source.getString(tempBasal.source) + " " + old.toString());
                         updateEarliestDataChange(oldDate);
                         updateEarliestDataChange(old.date);
                         scheduleTemporaryBasalChange();
@@ -1031,7 +894,8 @@ public class DatabaseHelper extends OrmLiteSqliteOpenHelper {
                             getDaoTemporaryBasal().delete(old); // need to delete/create because date may change too
                             old.copyFrom(tempBasal);
                             getDaoTemporaryBasal().create(old);
-                            log.debug("TEMPBASAL: Updating record by _id from: " + Source.getString(tempBasal.source) + " " + old.toString());
+                            if (L.isEnabled(L.DATABASE))
+                                log.debug("TEMPBASAL: Updating record by _id from: " + Source.getString(tempBasal.source) + " " + old.toString());
                             updateEarliestDataChange(oldDate);
                             updateEarliestDataChange(old.date);
                             scheduleTemporaryBasalChange();
@@ -1040,14 +904,16 @@ public class DatabaseHelper extends OrmLiteSqliteOpenHelper {
                     }
                 }
                 getDaoTemporaryBasal().create(tempBasal);
-                log.debug("TEMPBASAL: New record from: " + Source.getString(tempBasal.source) + " " + tempBasal.toString());
+                if (L.isEnabled(L.DATABASE))
+                    log.debug("TEMPBASAL: New record from: " + Source.getString(tempBasal.source) + " " + tempBasal.toString());
                 updateEarliestDataChange(tempBasal.date);
                 scheduleTemporaryBasalChange();
                 return true;
             }
             if (tempBasal.source == Source.USER) {
                 getDaoTemporaryBasal().create(tempBasal);
-                log.debug("TEMPBASAL: New record from: " + Source.getString(tempBasal.source) + " " + tempBasal.toString());
+                if (L.isEnabled(L.DATABASE))
+                    log.debug("TEMPBASAL: New record from: " + Source.getString(tempBasal.source) + " " + tempBasal.toString());
                 updateEarliestDataChange(tempBasal.date);
                 scheduleTemporaryBasalChange();
                 return true;
@@ -1087,7 +953,8 @@ public class DatabaseHelper extends OrmLiteSqliteOpenHelper {
     private static void scheduleTemporaryBasalChange() {
         class PostRunnable implements Runnable {
             public void run() {
-                log.debug("Firing EventTempBasalChange");
+                if (L.isEnabled(L.DATABASE))
+                    log.debug("Firing EventTempBasalChange");
                 MainApp.bus().post(new EventReloadTempBasalData());
                 MainApp.bus().post(new EventTempBasalChange());
                 if (earliestDataChange != null)
@@ -1125,15 +992,16 @@ public class DatabaseHelper extends OrmLiteSqliteOpenHelper {
     public void createTempBasalFromJsonIfNotExists(JSONObject trJson) {
         try {
             if (trJson.has("originalExtendedAmount")) { // extended bolus uploaded as temp basal
-                ExtendedBolus extendedBolus = new ExtendedBolus();
-                extendedBolus.source = Source.NIGHTSCOUT;
-                extendedBolus.date = trJson.getLong("mills");
-                extendedBolus.pumpId = trJson.has("pumpId") ? trJson.getLong("pumpId") : 0;
-                extendedBolus.durationInMinutes = trJson.getInt("duration");
-                extendedBolus.insulin = trJson.getDouble("originalExtendedAmount");
-                extendedBolus._id = trJson.getString("_id");
-                if (!VirtualPumpPlugin.getFakingStatus()) {
-                    VirtualPumpPlugin.setFakingStatus(true);
+                ExtendedBolus extendedBolus = new ExtendedBolus()
+                        .source(Source.NIGHTSCOUT)
+                        .date(trJson.getLong("mills"))
+                        .pumpId(trJson.has("pumpId") ? trJson.getLong("pumpId") : 0)
+                        .durationInMinutes(trJson.getInt("duration"))
+                        .insulin(trJson.getDouble("originalExtendedAmount"))
+                        ._id(trJson.getString("_id"));
+                // if faking found in NS, adapt AAPS to use it too
+                if (!VirtualPumpPlugin.getPlugin().getFakingStatus()) {
+                    VirtualPumpPlugin.getPlugin().setFakingStatus(true);
                     updateEarliestDataChange(0);
                     scheduleTemporaryBasalChange();
                 }
@@ -1146,17 +1014,18 @@ public class DatabaseHelper extends OrmLiteSqliteOpenHelper {
                 extendedBolus.durationInMinutes = 0;
                 extendedBolus.insulin = 0;
                 extendedBolus._id = trJson.getString("_id");
-                if (!VirtualPumpPlugin.getFakingStatus()) {
-                    VirtualPumpPlugin.setFakingStatus(true);
+                // if faking found in NS, adapt AAPS to use it too
+                if (!VirtualPumpPlugin.getPlugin().getFakingStatus()) {
+                    VirtualPumpPlugin.getPlugin().setFakingStatus(true);
                     updateEarliestDataChange(0);
                     scheduleTemporaryBasalChange();
                 }
                 createOrUpdate(extendedBolus);
             } else {
-                TemporaryBasal tempBasal = new TemporaryBasal();
-                tempBasal.date = trJson.getLong("mills");
-                tempBasal.source = Source.NIGHTSCOUT;
-                tempBasal.pumpId = trJson.has("pumpId") ? trJson.getLong("pumpId") : 0;
+                TemporaryBasal tempBasal = new TemporaryBasal()
+                        .date(trJson.getLong("mills"))
+                        .source(Source.NIGHTSCOUT)
+                        .pumpId(trJson.has("pumpId") ? trJson.getLong("pumpId") : 0);
                 if (trJson.has("duration")) {
                     tempBasal.durationInMinutes = trJson.getInt("duration");
                 }
@@ -1172,14 +1041,15 @@ public class DatabaseHelper extends OrmLiteSqliteOpenHelper {
                 createOrUpdate(tempBasal);
             }
         } catch (JSONException e) {
-            log.error("Unhandled exception", e);
+            log.error("Unhandled exception: " + trJson.toString(), e);
         }
     }
 
     public void deleteTempBasalById(String _id) {
         TemporaryBasal stored = findTempBasalById(_id);
         if (stored != null) {
-            log.debug("TEMPBASAL: Removing TempBasal record from database: " + stored.toString());
+            if (L.isEnabled(L.DATABASE))
+                log.debug("TEMPBASAL: Removing TempBasal record from database: " + stored.toString());
             delete(stored);
             updateEarliestDataChange(stored.date);
             scheduleTemporaryBasalChange();
@@ -1210,22 +1080,33 @@ public class DatabaseHelper extends OrmLiteSqliteOpenHelper {
 
     public boolean createOrUpdate(ExtendedBolus extendedBolus) {
         try {
+            if (L.isEnabled(L.DATABASE))
+                log.debug("EXTENDEDBOLUS: createOrUpdate: " + Source.getString(extendedBolus.source) + " " + extendedBolus.log());
+
             ExtendedBolus old;
             extendedBolus.date = roundDateToSec(extendedBolus.date);
 
             if (extendedBolus.source == Source.PUMP) {
-                // check for changed from pump change in NS
-                QueryBuilder<ExtendedBolus, Long> queryBuilder = getDaoExtendedBolus().queryBuilder();
-                Where where = queryBuilder.where();
-                where.eq("pumpId", extendedBolus.pumpId);
-                PreparedQuery<ExtendedBolus> preparedQuery = queryBuilder.prepare();
-                List<ExtendedBolus> trList = getDaoExtendedBolus().query(preparedQuery);
-                if (trList.size() > 0) {
-                    // do nothing, pump history record cannot be changed
-                    return false;
+                // if pumpId == 0 do not check for existing pumpId
+                // used with pumps without history
+                // and insight where record as added first without pumpId
+                // and then is record updated with pumpId
+                if (extendedBolus.pumpId == 0) {
+                    getDaoExtendedBolus().createOrUpdate(extendedBolus);
+                } else {
+                    QueryBuilder<ExtendedBolus, Long> queryBuilder = getDaoExtendedBolus().queryBuilder();
+                    Where where = queryBuilder.where();
+                    where.eq("pumpId", extendedBolus.pumpId);
+                    PreparedQuery<ExtendedBolus> preparedQuery = queryBuilder.prepare();
+                    List<ExtendedBolus> trList = getDaoExtendedBolus().query(preparedQuery);
+                    if (trList.size() > 1) {
+                        log.error("EXTENDEDBOLUS: Multiple records found for pumpId: " + extendedBolus.pumpId);
+                        return false;
+                    }
+                    getDaoExtendedBolus().createOrUpdate(extendedBolus);
                 }
-                getDaoExtendedBolus().create(extendedBolus);
-                log.debug("EXTENDEDBOLUS: New record from: " + Source.getString(extendedBolus.source) + " " + extendedBolus.toString());
+                if (L.isEnabled(L.DATABASE))
+                    log.debug("EXTENDEDBOLUS: New record from: " + Source.getString(extendedBolus.source) + " " + extendedBolus.log());
                 updateEarliestDataChange(extendedBolus.date);
                 scheduleExtendedBolusChange();
                 return true;
@@ -1238,7 +1119,8 @@ public class DatabaseHelper extends OrmLiteSqliteOpenHelper {
                         getDaoExtendedBolus().delete(old); // need to delete/create because date may change too
                         old.copyFrom(extendedBolus);
                         getDaoExtendedBolus().create(old);
-                        log.debug("EXTENDEDBOLUS: Updating record by date from: " + Source.getString(extendedBolus.source) + " " + old.toString());
+                        if (L.isEnabled(L.DATABASE))
+                            log.debug("EXTENDEDBOLUS: Updating record by date from: " + Source.getString(extendedBolus.source) + " " + old.log());
                         updateEarliestDataChange(oldDate);
                         updateEarliestDataChange(old.date);
                         scheduleExtendedBolusChange();
@@ -1260,7 +1142,8 @@ public class DatabaseHelper extends OrmLiteSqliteOpenHelper {
                             getDaoExtendedBolus().delete(old); // need to delete/create because date may change too
                             old.copyFrom(extendedBolus);
                             getDaoExtendedBolus().create(old);
-                            log.debug("EXTENDEDBOLUS: Updating record by _id from: " + Source.getString(extendedBolus.source) + " " + old.toString());
+                            if (L.isEnabled(L.DATABASE))
+                                log.debug("EXTENDEDBOLUS: Updating record by _id from: " + Source.getString(extendedBolus.source) + " " + old.log());
                             updateEarliestDataChange(oldDate);
                             updateEarliestDataChange(old.date);
                             scheduleExtendedBolusChange();
@@ -1269,14 +1152,16 @@ public class DatabaseHelper extends OrmLiteSqliteOpenHelper {
                     }
                 }
                 getDaoExtendedBolus().create(extendedBolus);
-                log.debug("EXTENDEDBOLUS: New record from: " + Source.getString(extendedBolus.source) + " " + extendedBolus.toString());
+                if (L.isEnabled(L.DATABASE))
+                    log.debug("EXTENDEDBOLUS: New record from: " + Source.getString(extendedBolus.source) + " " + extendedBolus.log());
                 updateEarliestDataChange(extendedBolus.date);
                 scheduleExtendedBolusChange();
                 return true;
             }
             if (extendedBolus.source == Source.USER) {
                 getDaoExtendedBolus().create(extendedBolus);
-                log.debug("EXTENDEDBOLUS: New record from: " + Source.getString(extendedBolus.source) + " " + extendedBolus.toString());
+                if (L.isEnabled(L.DATABASE))
+                    log.debug("EXTENDEDBOLUS: New record from: " + Source.getString(extendedBolus.source) + " " + extendedBolus.log());
                 updateEarliestDataChange(extendedBolus.date);
                 scheduleExtendedBolusChange();
                 return true;
@@ -1285,6 +1170,17 @@ public class DatabaseHelper extends OrmLiteSqliteOpenHelper {
             log.error("Unhandled exception", e);
         }
         return false;
+    }
+
+    public ExtendedBolus getExtendedBolusByPumpId(long pumpId) {
+        try {
+            return getDaoExtendedBolus().queryBuilder()
+                    .where().eq("pumpId", pumpId)
+                    .queryForFirst();
+        } catch (SQLException e) {
+            log.error("Unhandled exception", e);
+        }
+        return null;
     }
 
     public void delete(ExtendedBolus extendedBolus) {
@@ -1316,7 +1212,8 @@ public class DatabaseHelper extends OrmLiteSqliteOpenHelper {
     public void deleteExtendedBolusById(String _id) {
         ExtendedBolus stored = findExtendedBolusById(_id);
         if (stored != null) {
-            log.debug("EXTENDEDBOLUS: Removing ExtendedBolus record from database: " + stored.toString());
+            if (L.isEnabled(L.DATABASE))
+                log.debug("EXTENDEDBOLUS: Removing ExtendedBolus record from database: " + stored.toString());
             delete(stored);
             updateEarliestDataChange(stored.date);
             scheduleExtendedBolusChange();
@@ -1360,43 +1257,17 @@ public class DatabaseHelper extends OrmLiteSqliteOpenHelper {
 }
      */
 
-    public void createExtendedBolusFromJsonIfNotExists(JSONObject trJson) {
-        try {
-            QueryBuilder<ExtendedBolus, Long> queryBuilder = null;
-            queryBuilder = getDaoExtendedBolus().queryBuilder();
-            Where where = queryBuilder.where();
-            where.eq("_id", trJson.getString("_id")).or().eq("date", trJson.getLong("mills"));
-            PreparedQuery<ExtendedBolus> preparedQuery = queryBuilder.prepare();
-            List<ExtendedBolus> list = getDaoExtendedBolus().query(preparedQuery);
-            ExtendedBolus extendedBolus;
-            if (list.size() == 0) {
-                extendedBolus = new ExtendedBolus();
-                extendedBolus.source = Source.NIGHTSCOUT;
-                if (Config.logIncommingData)
-                    log.debug("Adding ExtendedBolus record to database: " + trJson.toString());
-                // Record does not exists. add
-            } else if (list.size() == 1) {
-                extendedBolus = list.get(0);
-                if (Config.logIncommingData)
-                    log.debug("Updating ExtendedBolus record in database: " + trJson.toString());
-            } else {
-                log.error("Something went wrong");
-                return;
-            }
-            extendedBolus.date = trJson.getLong("mills");
-            extendedBolus.durationInMinutes = trJson.has("duration") ? trJson.getInt("duration") : 0;
-            extendedBolus.insulin = trJson.getDouble("relative");
-            extendedBolus._id = trJson.getString("_id");
+    public void createExtendedBolusFromJsonIfNotExists(JSONObject json) {
+        ExtendedBolus extendedBolus = ExtendedBolus.createFromJson(json);
+        if (extendedBolus != null)
             createOrUpdate(extendedBolus);
-        } catch (SQLException | JSONException e) {
-            log.error("Unhandled exception", e);
-        }
     }
 
     private static void scheduleExtendedBolusChange() {
         class PostRunnable implements Runnable {
             public void run() {
-                log.debug("Firing EventExtendedBolusChange");
+                if (L.isEnabled(L.DATABASE))
+                    log.debug("Firing EventExtendedBolusChange");
                 MainApp.bus().post(new EventReloadTreatmentData(new EventExtendedBolusChange()));
                 if (earliestDataChange != null)
                     MainApp.bus().post(new EventNewHistoryData(earliestDataChange));
@@ -1436,6 +1307,15 @@ public class DatabaseHelper extends OrmLiteSqliteOpenHelper {
         scheduleCareportalEventChange();
     }
 
+    public CareportalEvent getCareportalEventFromTimestamp(long timestamp) {
+        try {
+            return getDaoCareportalEvents().queryForId(timestamp);
+        } catch (SQLException e) {
+            log.error("Unhandled exception", e);
+        }
+        return null;
+    }
+
     @Nullable
     public CareportalEvent getLastCareportalEvent(String event) {
         try {
@@ -1466,6 +1346,7 @@ public class DatabaseHelper extends OrmLiteSqliteOpenHelper {
             where.ge("date", mills);
             PreparedQuery<CareportalEvent> preparedQuery = queryBuilder.prepare();
             careportalEvents = getDaoCareportalEvents().query(preparedQuery);
+            preprocessOpenAPSOfflineEvents(careportalEvents);
             return careportalEvents;
         } catch (SQLException e) {
             log.error("Unhandled exception", e);
@@ -1473,13 +1354,41 @@ public class DatabaseHelper extends OrmLiteSqliteOpenHelper {
         return new ArrayList<>();
     }
 
-    public List<CareportalEvent> getCareportalEventsFromTime(boolean ascending) {
+    public void preprocessOpenAPSOfflineEvents(List<CareportalEvent> list) {
+        OverlappingIntervals offlineEvents = new OverlappingIntervals();
+        for (int i = 0; i < list.size(); i++) {
+            CareportalEvent event = list.get(i);
+            if (!event.eventType.equals(CareportalEvent.OPENAPSOFFLINE)) continue;
+            offlineEvents.add(event);
+        }
+
+    }
+
+    public List<CareportalEvent> getCareportalEventsFromTime(long mills, String type, boolean ascending) {
+        try {
+            List<CareportalEvent> careportalEvents;
+            QueryBuilder<CareportalEvent, Long> queryBuilder = getDaoCareportalEvents().queryBuilder();
+            queryBuilder.orderBy("date", ascending);
+            Where where = queryBuilder.where();
+            where.ge("date", mills).and().eq("eventType", type);
+            PreparedQuery<CareportalEvent> preparedQuery = queryBuilder.prepare();
+            careportalEvents = getDaoCareportalEvents().query(preparedQuery);
+            preprocessOpenAPSOfflineEvents(careportalEvents);
+            return careportalEvents;
+        } catch (SQLException e) {
+            log.error("Unhandled exception", e);
+        }
+        return new ArrayList<>();
+    }
+
+    public List<CareportalEvent> getCareportalEvents(boolean ascending) {
         try {
             List<CareportalEvent> careportalEvents;
             QueryBuilder<CareportalEvent, Long> queryBuilder = getDaoCareportalEvents().queryBuilder();
             queryBuilder.orderBy("date", ascending);
             PreparedQuery<CareportalEvent> preparedQuery = queryBuilder.prepare();
             careportalEvents = getDaoCareportalEvents().query(preparedQuery);
+            preprocessOpenAPSOfflineEvents(careportalEvents);
             return careportalEvents;
         } catch (SQLException e) {
             log.error("Unhandled exception", e);
@@ -1489,7 +1398,7 @@ public class DatabaseHelper extends OrmLiteSqliteOpenHelper {
 
     public void deleteCareportalEventById(String _id) {
         try {
-            QueryBuilder<CareportalEvent, Long> queryBuilder = null;
+            QueryBuilder<CareportalEvent, Long> queryBuilder;
             queryBuilder = getDaoCareportalEvents().queryBuilder();
             Where where = queryBuilder.where();
             where.eq("_id", _id);
@@ -1498,11 +1407,11 @@ public class DatabaseHelper extends OrmLiteSqliteOpenHelper {
 
             if (list.size() == 1) {
                 CareportalEvent record = list.get(0);
-                if (Config.logIncommingData)
-                    log.debug("Removing CareportalEvent record from database: " + record.log());
+                if (L.isEnabled(L.DATABASE))
+                    log.debug("Removing CareportalEvent record from database: " + record.toString());
                 delete(record);
             } else {
-                if (Config.logIncommingData)
+                if (L.isEnabled(L.DATABASE))
                     log.debug("CareportalEvent not found database: " + _id);
             }
         } catch (SQLException e) {
@@ -1512,7 +1421,7 @@ public class DatabaseHelper extends OrmLiteSqliteOpenHelper {
 
     public void createCareportalEventFromJsonIfNotExists(JSONObject trJson) {
         try {
-            QueryBuilder<CareportalEvent, Long> queryBuilder = null;
+            QueryBuilder<CareportalEvent, Long> queryBuilder;
             queryBuilder = getDaoCareportalEvents().queryBuilder();
             Where where = queryBuilder.where();
             where.eq("_id", trJson.getString("_id")).or().eq("date", trJson.getLong("mills"));
@@ -1522,12 +1431,12 @@ public class DatabaseHelper extends OrmLiteSqliteOpenHelper {
             if (list.size() == 0) {
                 careportalEvent = new CareportalEvent();
                 careportalEvent.source = Source.NIGHTSCOUT;
-                if (Config.logIncommingData)
+                if (L.isEnabled(L.DATABASE))
                     log.debug("Adding CareportalEvent record to database: " + trJson.toString());
                 // Record does not exists. add
             } else if (list.size() == 1) {
                 careportalEvent = list.get(0);
-                if (Config.logIncommingData)
+                if (L.isEnabled(L.DATABASE))
                     log.debug("Updating CareportalEvent record in database: " + trJson.toString());
             } else {
                 log.error("Something went wrong");
@@ -1539,14 +1448,15 @@ public class DatabaseHelper extends OrmLiteSqliteOpenHelper {
             careportalEvent._id = trJson.getString("_id");
             createOrUpdate(careportalEvent);
         } catch (SQLException | JSONException e) {
-            log.error("Unhandled exception", e);
+            log.error("Unhandled exception: " + trJson.toString(), e);
         }
     }
 
     private static void scheduleCareportalEventChange() {
         class PostRunnable implements Runnable {
             public void run() {
-                log.debug("Firing scheduleCareportalEventChange");
+                if (L.isEnabled(L.DATABASE))
+                    log.debug("Firing scheduleCareportalEventChange");
                 MainApp.bus().post(new EventCareportalEventChange());
                 scheduledCareportalEventPost = null;
             }
@@ -1569,14 +1479,32 @@ public class DatabaseHelper extends OrmLiteSqliteOpenHelper {
             List<ProfileSwitch> profileSwitches;
             QueryBuilder<ProfileSwitch, Long> queryBuilder = daoProfileSwitch.queryBuilder();
             queryBuilder.orderBy("date", ascending);
-            queryBuilder.limit(20L);
+            queryBuilder.limit(100L);
             PreparedQuery<ProfileSwitch> preparedQuery = queryBuilder.prepare();
             profileSwitches = daoProfileSwitch.query(preparedQuery);
             return profileSwitches;
         } catch (SQLException e) {
             log.error("Unhandled exception", e);
         }
-        return new ArrayList<ProfileSwitch>();
+        return new ArrayList<>();
+    }
+
+    public List<ProfileSwitch> getProfileSwitchEventsFromTime(long mills, boolean ascending) {
+        try {
+            Dao<ProfileSwitch, Long> daoProfileSwitch = getDaoProfileSwitch();
+            List<ProfileSwitch> profileSwitches;
+            QueryBuilder<ProfileSwitch, Long> queryBuilder = daoProfileSwitch.queryBuilder();
+            queryBuilder.orderBy("date", ascending);
+            queryBuilder.limit(100L);
+            Where where = queryBuilder.where();
+            where.ge("date", mills);
+            PreparedQuery<ProfileSwitch> preparedQuery = queryBuilder.prepare();
+            profileSwitches = daoProfileSwitch.query(preparedQuery);
+            return profileSwitches;
+        } catch (SQLException e) {
+            log.error("Unhandled exception", e);
+        }
+        return new ArrayList<>();
     }
 
     public boolean createOrUpdate(ProfileSwitch profileSwitch) {
@@ -1592,7 +1520,8 @@ public class DatabaseHelper extends OrmLiteSqliteOpenHelper {
                         profileSwitch.profileName = old.profileName; // preserver profileName to prevent multiple CPP extension
                         getDaoProfileSwitch().delete(old); // need to delete/create because date may change too
                         getDaoProfileSwitch().create(profileSwitch);
-                        log.debug("PROFILESWITCH: Updating record by date from: " + Source.getString(profileSwitch.source) + " " + old.toString());
+                        if (L.isEnabled(L.DATABASE))
+                            log.debug("PROFILESWITCH: Updating record by date from: " + Source.getString(profileSwitch.source) + " " + old.toString());
                         scheduleProfileSwitchChange();
                         return true;
                     }
@@ -1611,7 +1540,8 @@ public class DatabaseHelper extends OrmLiteSqliteOpenHelper {
                             getDaoProfileSwitch().delete(old); // need to delete/create because date may change too
                             old.copyFrom(profileSwitch);
                             getDaoProfileSwitch().create(old);
-                            log.debug("PROFILESWITCH: Updating record by _id from: " + Source.getString(profileSwitch.source) + " " + old.toString());
+                            if (L.isEnabled(L.DATABASE))
+                                log.debug("PROFILESWITCH: Updating record by _id from: " + Source.getString(profileSwitch.source) + " " + old.toString());
                             scheduleProfileSwitchChange();
                             return true;
                         }
@@ -1620,13 +1550,15 @@ public class DatabaseHelper extends OrmLiteSqliteOpenHelper {
                 // look for already added percentage from NS
                 profileSwitch.profileName = PercentageSplitter.pureName(profileSwitch.profileName);
                 getDaoProfileSwitch().create(profileSwitch);
-                log.debug("PROFILESWITCH: New record from: " + Source.getString(profileSwitch.source) + " " + profileSwitch.toString());
+                if (L.isEnabled(L.DATABASE))
+                    log.debug("PROFILESWITCH: New record from: " + Source.getString(profileSwitch.source) + " " + profileSwitch.toString());
                 scheduleProfileSwitchChange();
                 return true;
             }
             if (profileSwitch.source == Source.USER) {
                 getDaoProfileSwitch().create(profileSwitch);
-                log.debug("PROFILESWITCH: New record from: " + Source.getString(profileSwitch.source) + " " + profileSwitch.toString());
+                if (L.isEnabled(L.DATABASE))
+                    log.debug("PROFILESWITCH: New record from: " + Source.getString(profileSwitch.source) + " " + profileSwitch.toString());
                 scheduleProfileSwitchChange();
                 return true;
             }
@@ -1648,9 +1580,10 @@ public class DatabaseHelper extends OrmLiteSqliteOpenHelper {
     private static void scheduleProfileSwitchChange() {
         class PostRunnable implements Runnable {
             public void run() {
-                log.debug("Firing EventProfileSwitchChange");
+                if (L.isEnabled(L.DATABASE))
+                    log.debug("Firing EventProfileNeedsUpdate");
                 MainApp.bus().post(new EventReloadProfileSwitchData());
-                MainApp.bus().post(new EventProfileSwitchChange());
+                MainApp.bus().post(new EventProfileNeedsUpdate());
                 scheduledProfileSwitchEventPost = null;
             }
         }
@@ -1693,15 +1626,30 @@ public class DatabaseHelper extends OrmLiteSqliteOpenHelper {
             if (trJson.has("profileJson"))
                 profileSwitch.profileJson = trJson.getString("profileJson");
             else {
-                ProfileStore store = ConfigBuilderPlugin.getActiveProfileInterface().getProfile();
-                Profile profile = store.getSpecificProfile(profileSwitch.profileName);
-                if (profile != null) {
-                    profileSwitch.profileJson = profile.getData().toString();
-                    log.debug("Profile switch prefilled with JSON from local store");
-                    // Update data in NS
-                    NSUpload.updateProfileSwitch(profileSwitch);
+                ProfileInterface profileInterface = ConfigBuilderPlugin.getPlugin().getActiveProfileInterface();
+                if (profileInterface != null) {
+                    ProfileStore store = profileInterface.getProfile();
+                    if (store != null) {
+                        Profile profile = store.getSpecificProfile(profileSwitch.profileName);
+                        if (profile != null) {
+                            profileSwitch.profileJson = profile.getData().toString();
+                            if (L.isEnabled(L.DATABASE))
+                                log.debug("Profile switch prefilled with JSON from local store");
+                            // Update data in NS
+                            NSUpload.updateProfileSwitch(profileSwitch);
+                        } else {
+                            if (L.isEnabled(L.DATABASE))
+                                log.debug("JSON for profile switch doesn't exist. Ignoring: " + trJson.toString());
+                            return;
+                        }
+                    } else {
+                        if (L.isEnabled(L.DATABASE))
+                            log.debug("Store for profile switch doesn't exist. Ignoring: " + trJson.toString());
+                        return;
+                    }
                 } else {
-                    log.debug("JSON for profile switch doesn't exist. Ignoring: " + trJson.toString());
+                    if (L.isEnabled(L.DATABASE))
+                        log.debug("No active profile interface. Ignoring: " + trJson.toString());
                     return;
                 }
             }
@@ -1709,14 +1657,15 @@ public class DatabaseHelper extends OrmLiteSqliteOpenHelper {
                 profileSwitch.profilePlugin = trJson.getString("profilePlugin");
             createOrUpdate(profileSwitch);
         } catch (JSONException e) {
-            log.error("Unhandled exception", e);
+            log.error("Unhandled exception: " + trJson.toString(), e);
         }
     }
 
     public void deleteProfileSwitchById(String _id) {
         ProfileSwitch stored = findProfileSwitchById(_id);
         if (stored != null) {
-            log.debug("PROFILESWITCH: Removing ProfileSwitch record from database: " + stored.toString());
+            if (L.isEnabled(L.DATABASE))
+                log.debug("PROFILESWITCH: Removing ProfileSwitch record from database: " + stored.toString());
             delete(stored);
             scheduleTemporaryTargetChange();
         }
@@ -1735,6 +1684,68 @@ public class DatabaseHelper extends OrmLiteSqliteOpenHelper {
             } else {
                 return null;
             }
+        } catch (SQLException e) {
+            log.error("Unhandled exception", e);
+        }
+        return null;
+    }
+
+    // ---------------- Insight history handling ---------------
+
+    public void createOrUpdate(InsightHistoryOffset offset) {
+        try {
+            getDaoInsightHistoryOffset().createOrUpdate(offset);
+        } catch (SQLException e) {
+            log.error("Unhandled exception", e);
+        }
+    }
+
+    public InsightHistoryOffset getInsightHistoryOffset(String pumpSerial) {
+        try {
+            return getDaoInsightHistoryOffset().queryForId(pumpSerial);
+        } catch (SQLException e) {
+            log.error("Unhandled exception", e);
+        }
+        return null;
+    }
+
+    public void createOrUpdate(InsightBolusID bolusID) {
+        try {
+            getDaoInsightBolusID().createOrUpdate(bolusID);
+        } catch (SQLException e) {
+            log.error("Unhandled exception", e);
+        }
+    }
+
+    public InsightBolusID getInsightBolusID(String pumpSerial, int bolusID, long timestamp) {
+        try {
+            return getDaoInsightBolusID().queryBuilder()
+                    .where().eq("pumpSerial", pumpSerial)
+                    .and().eq("bolusID", bolusID)
+                    .and().between("timestamp", timestamp - 259200000, timestamp + 259200000)
+                    .queryForFirst();
+        } catch (SQLException e) {
+            log.error("Unhandled exception", e);
+        }
+        return null;
+    }
+
+    public void createOrUpdate(InsightPumpID pumpID) {
+        try {
+            getDaoInsightPumpID().createOrUpdate(pumpID);
+        } catch (SQLException e) {
+            log.error("Unhandled exception", e);
+        }
+    }
+
+    public InsightPumpID getPumpStoppedEvent(String pumpSerial, long before) {
+        try {
+            return getDaoInsightPumpID().queryBuilder()
+                    .orderBy("timestamp", false)
+                    .where().eq("pumpSerial", pumpSerial)
+                    .and().in("eventType", "PumpStopped", "PumpPaused")
+                    .and().lt("timestamp", before)
+                    .queryForFirst();
         } catch (SQLException e) {
             log.error("Unhandled exception", e);
         }
