@@ -9,25 +9,26 @@ import android.graphics.Color;
 import com.j256.ormlite.field.DatabaseField;
 import com.j256.ormlite.table.DatabaseTable;
 
+import org.json.JSONObject;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import java.util.Date;
 import java.util.Objects;
 
 import info.nightscout.androidaps.Constants;
-import info.nightscout.androidaps.MainApp;
 import info.nightscout.androidaps.data.Iob;
 import info.nightscout.androidaps.data.IobTotal;
-import info.nightscout.androidaps.data.Profile;
 import info.nightscout.androidaps.interfaces.InsulinInterface;
 import info.nightscout.androidaps.interfaces.Interval;
-import info.nightscout.androidaps.plugins.ConfigBuilder.ConfigBuilderPlugin;
-import info.nightscout.androidaps.plugins.Overview.graphExtensions.DataPointWithLabelInterface;
-import info.nightscout.androidaps.plugins.Overview.graphExtensions.PointsWithLabelGraphSeries;
-import info.nightscout.utils.DateUtil;
-import info.nightscout.utils.DecimalFormatter;
-import info.nightscout.utils.Round;
+import info.nightscout.androidaps.logging.L;
+import info.nightscout.androidaps.plugins.configBuilder.ConfigBuilderPlugin;
+import info.nightscout.androidaps.plugins.general.overview.graphExtensions.DataPointWithLabelInterface;
+import info.nightscout.androidaps.plugins.general.overview.graphExtensions.PointsWithLabelGraphSeries;
+import info.nightscout.androidaps.plugins.treatments.Treatment;
+import info.nightscout.androidaps.utils.DateUtil;
+import info.nightscout.androidaps.utils.DecimalFormatter;
+import info.nightscout.androidaps.utils.JsonHelper;
+import info.nightscout.androidaps.utils.Round;
 
 /**
  * Created by mike on 21.05.2017.
@@ -35,7 +36,7 @@ import info.nightscout.utils.Round;
 
 @DatabaseTable(tableName = DatabaseHelper.DATABASE_EXTENDEDBOLUSES)
 public class ExtendedBolus implements Interval, DataPointWithLabelInterface {
-    private static Logger log = LoggerFactory.getLogger(ExtendedBolus.class);
+    private static Logger log = LoggerFactory.getLogger(L.DATABASE);
 
     @DatabaseField(id = true)
     public long date;
@@ -57,7 +58,7 @@ public class ExtendedBolus implements Interval, DataPointWithLabelInterface {
     public int durationInMinutes = 0; // duration == 0 means end of extended bolus
 
     @DatabaseField
-    public int insulinInterfaceID = InsulinInterface.FASTACTINGINSULIN;
+    public int insulinInterfaceID = InsulinInterface.OREF_RAPID_ACTING;
     @DatabaseField
     public double dia = Constants.defaultDIA;
 
@@ -66,6 +67,36 @@ public class ExtendedBolus implements Interval, DataPointWithLabelInterface {
 
     public ExtendedBolus(long date) {
         this.date = date;
+    }
+
+    public ExtendedBolus date(long date) {
+        this.date = date;
+        return this;
+    }
+
+    public ExtendedBolus insulin(double insulin) {
+        this.insulin = insulin;
+        return this;
+    }
+
+    public ExtendedBolus pumpId(long pumpId) {
+        this.pumpId = pumpId;
+        return this;
+    }
+
+    public ExtendedBolus source(int source) {
+        this.source = source;
+        return this;
+    }
+
+    public ExtendedBolus durationInMinutes(int durationInMinutes) {
+        this.durationInMinutes = durationInMinutes;
+        return this;
+    }
+
+    public ExtendedBolus _id(String _id) {
+        this._id = _id;
+        return this;
     }
 
     public boolean isEqual(ExtendedBolus other) {
@@ -91,6 +122,16 @@ public class ExtendedBolus implements Interval, DataPointWithLabelInterface {
         pumpId = t.pumpId;
     }
 
+    public static ExtendedBolus createFromJson(JSONObject json) {
+        ExtendedBolus extendedBolus = new ExtendedBolus()
+                .source(Source.NIGHTSCOUT)
+                .date(JsonHelper.safeGetLong(json, "mills"))
+                .durationInMinutes(JsonHelper.safeGetInt(json, "duration"))
+                .insulin(JsonHelper.safeGetDouble(json, "relative") / 60 * JsonHelper.safeGetInt(json, "duration"))
+                ._id(JsonHelper.safeGetString(json, "_id"))
+                .pumpId(JsonHelper.safeGetLong(json, "pumpId"));
+        return extendedBolus;
+    }
     // -------- Interval interface ---------
 
     Long cuttedEnd = null;
@@ -155,7 +196,7 @@ public class ExtendedBolus implements Interval, DataPointWithLabelInterface {
     // -------- Interval interface end ---------
 
     public String log() {
-        return "Bolus{" +
+        return "ExtendedBolus{" +
                 "date= " + date +
                 ", date= " + DateUtil.dateAndTimeString(date) +
                 ", isValid=" + isValid +
@@ -176,7 +217,7 @@ public class ExtendedBolus implements Interval, DataPointWithLabelInterface {
 
     public IobTotal iobCalc(long time) {
         IobTotal result = new IobTotal(time);
-        InsulinInterface insulinInterface = ConfigBuilderPlugin.getActiveInsulin();
+        InsulinInterface insulinInterface = ConfigBuilderPlugin.getPlugin().getActiveInsulin();
 
         int realDuration = getDurationToTime(time);
 
@@ -285,4 +326,5 @@ public class ExtendedBolus implements Interval, DataPointWithLabelInterface {
     public int getColor() {
         return Color.CYAN;
     }
+
 }
