@@ -10,9 +10,10 @@ import info.nightscout.androidaps.R;
 public class EventPumpStatusChanged extends Event {
     public static final int CONNECTING = 0;
     public static final int CONNECTED = 1;
-    public static final int PERFORMING = 2;
-    public static final int DISCONNECTING = 3;
-    public static final int DISCONNECTED = 4;
+    public static final int HANDSHAKING = 2;
+    public static final int PERFORMING = 3;
+    public static final int DISCONNECTING = 4;
+    public static final int DISCONNECTED = 5;
 
     public int sStatus = DISCONNECTED;
     public int sSecondsElapsed = 0;
@@ -46,13 +47,15 @@ public class EventPumpStatusChanged extends Event {
 
     public String textStatus() {
         if (sStatus == CONNECTING)
-            return String.format(MainApp.sResources.getString(R.string.danar_history_connectingfor), sSecondsElapsed);
+            return String.format(MainApp.gs(R.string.danar_history_connectingfor), sSecondsElapsed);
+        else if (sStatus == HANDSHAKING)
+            return MainApp.gs(R.string.handshaking);
         else if (sStatus == CONNECTED)
-            return MainApp.sResources.getString(R.string.connected);
+            return MainApp.gs(R.string.connected);
         else if (sStatus == PERFORMING)
             return sPerfomingAction;
         else if (sStatus == DISCONNECTING)
-            return MainApp.sResources.getString(R.string.disconnecting);
+            return MainApp.gs(R.string.disconnecting);
         else if (sStatus == DISCONNECTED)
             return "";
         return "";
