@@ -1,15 +1,13 @@
 package info.nightscout.androidaps.interfaces;
 
-import java.util.Date;
-
-import info.nightscout.androidaps.plugins.Loop.APSResult;
+import info.nightscout.androidaps.plugins.aps.loop.APSResult;
 
 /**
  * Created by mike on 10.06.2016.
  */
 public interface APSInterface {
     public APSResult getLastAPSResult();
-    public Date getLastAPSRun();
+    public long getLastAPSRun();
 
-    public void invoke(String initiator);
+    public void invoke(String initiator, boolean tempBasalFallback);
 }
