@@ -5,11 +5,12 @@ import java.util.List;
 import info.nightscout.androidaps.data.DetailedBolusInfo;
 import info.nightscout.androidaps.data.IobTotal;
 import info.nightscout.androidaps.data.MealData;
+import info.nightscout.androidaps.data.Profile;
 import info.nightscout.androidaps.db.ExtendedBolus;
 import info.nightscout.androidaps.db.ProfileSwitch;
 import info.nightscout.androidaps.db.TempTarget;
 import info.nightscout.androidaps.db.TemporaryBasal;
-import info.nightscout.androidaps.db.Treatment;
+import info.nightscout.androidaps.plugins.treatments.Treatment;
 import info.nightscout.androidaps.data.Intervals;
 import info.nightscout.androidaps.data.ProfileIntervals;
 
@@ -24,12 +25,13 @@ public interface TreatmentsInterface {
     IobTotal getLastCalculationTreatments();
     IobTotal getCalculationToTimeTreatments(long time);
     IobTotal getLastCalculationTempBasals();
-    IobTotal getCalculationToTimeTempBasals(long time);
+    IobTotal getCalculationToTimeTempBasals(long time, Profile profile);
 
     MealData getMealData();
 
     List<Treatment> getTreatmentsFromHistory();
     List<Treatment> getTreatments5MinBackFromHistory(long time);
+    long getLastBolusTime();
 
     // real basals (not faked by extended bolus)
     boolean isInHistoryRealTempBasalInProgress();
@@ -40,8 +42,6 @@ public interface TreatmentsInterface {
     // basal that can be faked by extended boluses
     boolean isTempBasalInProgress();
     TemporaryBasal getTempBasalFromHistory(long time);
-    double getTempBasalAbsoluteRateHistory();
-    double getTempBasalRemainingMinutesFromHistory();
     Intervals<TemporaryBasal> getTemporaryBasalsFromHistory();
 
     boolean isInHistoryExtendedBoluslInProgress();
@@ -50,11 +50,12 @@ public interface TreatmentsInterface {
 
     boolean addToHistoryExtendedBolus(ExtendedBolus extendedBolus);
 
-    boolean addToHistoryTreatment(DetailedBolusInfo detailedBolusInfo);
+    boolean addToHistoryTreatment(DetailedBolusInfo detailedBolusInfo, boolean allowUpdate);
 
     TempTarget getTempTargetFromHistory();
     TempTarget getTempTargetFromHistory(long time);
     Intervals<TempTarget> getTempTargetsFromHistory();
+    void addToHistoryTempTarget(TempTarget tempTarget);
 
     ProfileSwitch getProfileSwitchFromHistory(long time);
     ProfileIntervals<ProfileSwitch> getProfileSwitchesFromHistory();
