@@ -11,62 +11,75 @@ import org.slf4j.LoggerFactory;
 import info.nightscout.androidaps.Constants;
 import info.nightscout.androidaps.MainApp;
 import info.nightscout.androidaps.R;
-import info.nightscout.androidaps.data.PumpEnactResult;
 import info.nightscout.androidaps.events.EventPumpStatusChanged;
 import info.nightscout.androidaps.interfaces.PumpInterface;
-import info.nightscout.androidaps.plugins.ConfigBuilder.ConfigBuilderPlugin;
-import info.nightscout.androidaps.plugins.Overview.events.EventDismissBolusprogressIfRunning;
-import info.nightscout.androidaps.plugins.Overview.events.EventNewNotification;
-import info.nightscout.androidaps.plugins.Overview.notifications.Notification;
+import info.nightscout.androidaps.logging.L;
+import info.nightscout.androidaps.plugins.configBuilder.ConfigBuilderPlugin;
+import info.nightscout.androidaps.plugins.general.overview.events.EventDismissBolusprogressIfRunning;
 import info.nightscout.androidaps.queue.events.EventQueueChanged;
-import info.nightscout.utils.SP;
+import info.nightscout.androidaps.utils.SP;
+import info.nightscout.androidaps.utils.T;
 
 /**
  * Created by mike on 09.11.2017.
  */
 
 public class QueueThread extends Thread {
-    private static Logger log = LoggerFactory.getLogger(QueueThread.class);
+    private Logger log = LoggerFactory.getLogger(L.PUMPQUEUE);
 
-    CommandQueue queue;
+    private CommandQueue queue;
 
-    private long connectionStartTime = 0;
-    private long lastCommandTime = 0;
     private boolean connectLogged = false;
+    boolean waitingForDisconnect = false;
 
     private PowerManager.WakeLock mWakeLock;
 
-    public QueueThread(CommandQueue queue) {
+    QueueThread(CommandQueue queue) {
         super();
 
         this.queue = queue;
-        PowerManager powerManager = (PowerManager) MainApp.instance().getApplicationContext().getSystemService(Context.POWER_SERVICE);
-        mWakeLock = powerManager.newWakeLock(PowerManager.SCREEN_DIM_WAKE_LOCK, "QueueThread");
+        Context context = MainApp.instance().getApplicationContext();
+        if (context != null) {
+            PowerManager powerManager = (PowerManager) context.getSystemService(Context.POWER_SERVICE);
+            if (powerManager != null)
+                mWakeLock = powerManager.newWakeLock(PowerManager.PARTIAL_WAKE_LOCK, "AndroidAPS:QueueThread");
+        }
     }
 
     @Override
     public final void run() {
-        mWakeLock.acquire();
+        if (mWakeLock != null)
+            mWakeLock.acquire(T.mins(10).msecs());
         MainApp.bus().post(new EventQueueChanged());
-        connectionStartTime = lastCommandTime = System.currentTimeMillis();
+        long lastCommandTime;
+        long connectionStartTime = lastCommandTime = System.currentTimeMillis();
 
         try {
             while (true) {
-                PumpInterface pump = ConfigBuilderPlugin.getActivePump();
+                PumpInterface pump = ConfigBuilderPlugin.getPlugin().getActivePump();
+                if (pump == null) {
+                    if (L.isEnabled(L.PUMPQUEUE))
+                        log.debug("pump == null");
+                    MainApp.bus().post(new EventPumpStatusChanged(MainApp.gs(R.string.pumpNotInitialized)));
+                    SystemClock.sleep(1000);
+                    continue;
+                }
                 long secondsElapsed = (System.currentTimeMillis() - connectionStartTime) / 1000;
 
                 if (!pump.isConnected() && secondsElapsed > Constants.PUMP_MAX_CONNECTION_TIME_IN_SECONDS) {
                     MainApp.bus().post(new EventDismissBolusprogressIfRunning(null));
-                    MainApp.bus().post(new EventPumpStatusChanged(MainApp.sResources.getString(R.string.connectiontimedout)));
-                    log.debug("QUEUE: timed out");
+                    MainApp.bus().post(new EventPumpStatusChanged(MainApp.gs(R.string.connectiontimedout)));
+                    if (L.isEnabled(L.PUMPQUEUE))
+                        log.debug("timed out");
                     pump.stopConnecting();
 
                     //BLUETOOTH-WATCHDOG
                     boolean watchdog = SP.getBoolean(R.string.key_btwatchdog, false);
-                    long last_watchdog = SP.getLong(R.string.key_btwatchdog_lastbark, 0l);
+                    long last_watchdog = SP.getLong(R.string.key_btwatchdog_lastbark, 0L);
                     watchdog = watchdog && System.currentTimeMillis() - last_watchdog > (Constants.MIN_WATCHDOG_INTERVAL_IN_SECONDS * 1000);
-                    if(watchdog) {
-                        log.debug("BT watchdog - toggeling the phonest bluetooth");
+                    if (watchdog) {
+                        if (L.isEnabled(L.PUMPQUEUE))
+                            log.debug("BT watchdog - toggeling the phonest bluetooth");
                         //write time
                         SP.putLong(R.string.key_btwatchdog_lastbark, System.currentTimeMillis());
                         //toggle BT
@@ -85,7 +98,8 @@ public class QueueThread extends Thread {
                         pump.connect("watchdog");
                     } else {
                         queue.clear();
-                        log.debug("QUEUE: no connection possible");
+                        if (L.isEnabled(L.PUMPQUEUE))
+                            log.debug("no connection possible");
                         MainApp.bus().post(new EventPumpStatusChanged(EventPumpStatusChanged.DISCONNECTING));
                         pump.disconnect("Queue empty");
                         MainApp.bus().post(new EventPumpStatusChanged(EventPumpStatusChanged.DISCONNECTED));
@@ -93,16 +107,25 @@ public class QueueThread extends Thread {
                     }
                 }
 
+                if (pump.isHandshakeInProgress()) {
+                    if (L.isEnabled(L.PUMPQUEUE))
+                        log.debug("handshaking " + secondsElapsed);
+                    MainApp.bus().post(new EventPumpStatusChanged(EventPumpStatusChanged.HANDSHAKING, (int) secondsElapsed));
+                    SystemClock.sleep(100);
+                    continue;
+                }
+
                 if (pump.isConnecting()) {
-                    log.debug("QUEUE: connecting " + secondsElapsed);
+                    if (L.isEnabled(L.PUMPQUEUE))
+                        log.debug("connecting " + secondsElapsed);
                     MainApp.bus().post(new EventPumpStatusChanged(EventPumpStatusChanged.CONNECTING, (int) secondsElapsed));
                     SystemClock.sleep(1000);
                     continue;
                 }
 
-
                 if (!pump.isConnected()) {
-                    log.debug("QUEUE: connect");
+                    if (L.isEnabled(L.PUMPQUEUE))
+                        log.debug("connect");
                     MainApp.bus().post(new EventPumpStatusChanged(EventPumpStatusChanged.CONNECTING, (int) secondsElapsed));
                     pump.connect("Connection needed");
                     SystemClock.sleep(1000);
@@ -112,12 +135,14 @@ public class QueueThread extends Thread {
                 if (queue.performing() == null) {
                     if (!connectLogged) {
                         connectLogged = true;
-                        log.debug("QUEUE: connection time " + secondsElapsed + "s");
+                        if (L.isEnabled(L.PUMPQUEUE))
+                            log.debug("connection time " + secondsElapsed + "s");
                     }
                     // Pickup 1st command and set performing variable
                     if (queue.size() > 0) {
                         queue.pickup();
-                        log.debug("QUEUE: performing " + queue.performing().status());
+                        if (L.isEnabled(L.PUMPQUEUE))
+                            log.debug("performing " + queue.performing().status());
                         MainApp.bus().post(new EventQueueChanged());
                         queue.performing().execute();
                         queue.resetPerforming();
@@ -131,21 +156,27 @@ public class QueueThread extends Thread {
                 if (queue.size() == 0 && queue.performing() == null) {
                     long secondsFromLastCommand = (System.currentTimeMillis() - lastCommandTime) / 1000;
                     if (secondsFromLastCommand >= 5) {
-                        log.debug("QUEUE: queue empty. disconnect");
+                        waitingForDisconnect = true;
+                        if (L.isEnabled(L.PUMPQUEUE))
+                            log.debug("queue empty. disconnect");
                         MainApp.bus().post(new EventPumpStatusChanged(EventPumpStatusChanged.DISCONNECTING));
                         pump.disconnect("Queue empty");
                         MainApp.bus().post(new EventPumpStatusChanged(EventPumpStatusChanged.DISCONNECTED));
+                        if (L.isEnabled(L.PUMPQUEUE))
+                            log.debug("disconnected");
                         return;
                     } else {
-                        log.debug("QUEUE: waiting for disconnect");
+                        if (L.isEnabled(L.PUMPQUEUE))
+                            log.debug("waiting for disconnect");
                         SystemClock.sleep(1000);
                     }
                 }
             }
         } finally {
-            mWakeLock.release();
+            if (mWakeLock != null)
+                mWakeLock.release();
+            if (L.isEnabled(L.PUMPQUEUE))
+                log.debug("thread end");
         }
     }
-
-
 }
