@@ -3,10 +3,9 @@ package info.nightscout.androidaps.queue.commands;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import info.nightscout.androidaps.Config;
-import info.nightscout.androidaps.MainApp;
 import info.nightscout.androidaps.data.PumpEnactResult;
-import info.nightscout.androidaps.plugins.ConfigBuilder.ConfigBuilderPlugin;
+import info.nightscout.androidaps.logging.L;
+import info.nightscout.androidaps.plugins.configBuilder.ConfigBuilderPlugin;
 import info.nightscout.androidaps.queue.Callback;
 
 /**
@@ -14,7 +13,7 @@ import info.nightscout.androidaps.queue.Callback;
  */
 
 public class CommandCancelExtendedBolus extends Command {
-    private static Logger log = LoggerFactory.getLogger(CommandCancelExtendedBolus.class);
+    private Logger log = LoggerFactory.getLogger(L.PUMPQUEUE);
 
     public CommandCancelExtendedBolus(Callback callback) {
         commandType = CommandType.EXTENDEDBOLUS;
@@ -23,9 +22,9 @@ public class CommandCancelExtendedBolus extends Command {
 
     @Override
     public void execute() {
-        PumpEnactResult r = ConfigBuilderPlugin.getActivePump().cancelExtendedBolus();
-        if (Config.logCongigBuilderActions)
-            log.debug("cancelExtendedBolus success: " + r.success + " enacted: " + r.enacted);
+        PumpEnactResult r = ConfigBuilderPlugin.getPlugin().getActivePump().cancelExtendedBolus();
+        if (L.isEnabled(L.PUMPQUEUE))
+            log.debug("Result success: " + r.success + " enacted: " + r.enacted);
         if (callback != null)
             callback.result(r).run();
     }
