@@ -1,9 +1,13 @@
 package info.nightscout.androidaps.queue.commands;
 
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
 import info.nightscout.androidaps.data.PumpEnactResult;
 import info.nightscout.androidaps.interfaces.DanaRInterface;
 import info.nightscout.androidaps.interfaces.PumpInterface;
-import info.nightscout.androidaps.plugins.ConfigBuilder.ConfigBuilderPlugin;
+import info.nightscout.androidaps.logging.L;
+import info.nightscout.androidaps.plugins.configBuilder.ConfigBuilderPlugin;
 import info.nightscout.androidaps.queue.Callback;
 
 /**
@@ -11,6 +15,8 @@ import info.nightscout.androidaps.queue.Callback;
  */
 
 public class CommandLoadEvents extends Command {
+    private Logger log = LoggerFactory.getLogger(L.PUMPQUEUE);
+
     public CommandLoadEvents(Callback callback) {
         commandType = CommandType.LOADEVENTS;
         this.callback = callback;
@@ -18,10 +24,12 @@ public class CommandLoadEvents extends Command {
 
     @Override
     public void execute() {
-        PumpInterface pump = ConfigBuilderPlugin.getActivePump();
+        PumpInterface pump = ConfigBuilderPlugin.getPlugin().getActivePump();
         if (pump instanceof DanaRInterface) {
             DanaRInterface danaPump = (DanaRInterface) pump;
             PumpEnactResult r = danaPump.loadEvents();
+            if (L.isEnabled(L.PUMPQUEUE))
+                log.debug("Result success: " + r.success + " enacted: " + r.enacted);
             if (callback != null)
                 callback.result(r).run();
         }
