@@ -1,18 +1,23 @@
 package info.nightscout.androidaps.queue.commands;
 
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
 import info.nightscout.androidaps.data.PumpEnactResult;
 import info.nightscout.androidaps.interfaces.DanaRInterface;
 import info.nightscout.androidaps.interfaces.PumpInterface;
-import info.nightscout.androidaps.plugins.ConfigBuilder.ConfigBuilderPlugin;
+import info.nightscout.androidaps.logging.L;
+import info.nightscout.androidaps.plugins.configBuilder.ConfigBuilderPlugin;
 import info.nightscout.androidaps.queue.Callback;
-import info.nightscout.androidaps.queue.commands.Command;
 
 /**
  * Created by mike on 10.11.2017.
  */
 
 public class CommandLoadHistory extends Command {
-    byte type;
+    private Logger log = LoggerFactory.getLogger(L.PUMPQUEUE);
+
+    private byte type;
 
     public CommandLoadHistory(byte type, Callback callback) {
         commandType = CommandType.LOADHISTORY;
@@ -22,10 +27,12 @@ public class CommandLoadHistory extends Command {
 
     @Override
     public void execute() {
-        PumpInterface pump = ConfigBuilderPlugin.getActivePump();
+        PumpInterface pump = ConfigBuilderPlugin.getPlugin().getActivePump();
         if (pump instanceof DanaRInterface) {
             DanaRInterface danaPump = (DanaRInterface) pump;
             PumpEnactResult r = danaPump.loadHistory(type);
+            if (L.isEnabled(L.PUMPQUEUE))
+                log.debug("Result success: " + r.success + " enacted: " + r.enacted);
             if (callback != null)
                 callback.result(r).run();
         }
