@@ -1,9 +1,5 @@
 package info.nightscout.androidaps.receivers;
 
-/**
- * Created by mike on 07.07.2016.
- */
-
 import android.app.AlarmManager;
 import android.app.PendingIntent;
 import android.content.BroadcastReceiver;
@@ -14,19 +10,29 @@ import android.os.PowerManager;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import java.util.Date;
-
 import info.nightscout.androidaps.Constants;
 import info.nightscout.androidaps.MainApp;
-import info.nightscout.androidaps.interfaces.PumpInterface;
 import info.nightscout.androidaps.data.Profile;
-import info.nightscout.androidaps.plugins.ConfigBuilder.ConfigBuilderPlugin;
-import info.nightscout.utils.LocalAlertUtils;
-import info.nightscout.utils.SP;
+import info.nightscout.androidaps.events.EventProfileNeedsUpdate;
+import info.nightscout.androidaps.interfaces.PumpInterface;
+import info.nightscout.androidaps.logging.L;
+import info.nightscout.androidaps.plugins.configBuilder.ConfigBuilderPlugin;
+import info.nightscout.androidaps.plugins.configBuilder.ProfileFunctions;
+import info.nightscout.androidaps.queue.commands.Command;
+import info.nightscout.androidaps.utils.DateUtil;
+import info.nightscout.androidaps.utils.FabricPrivacy;
+import info.nightscout.androidaps.utils.LocalAlertUtils;
+import info.nightscout.androidaps.utils.T;
 
+
+/**
+ * Created by mike on 07.07.2016.
+ */
 public class KeepAliveReceiver extends BroadcastReceiver {
-    private static Logger log = LoggerFactory.getLogger(KeepAliveReceiver.class);
-    public static final long STATUS_UPDATE_FREQUENCY = 15 * 60 * 1000L;
+    private static Logger log = LoggerFactory.getLogger(L.CORE);
+    public static final long STATUS_UPDATE_FREQUENCY = T.mins(15).msecs();
+    private static long lastReadStatus = 0;
+    private static long lastRun = 0;
 
     public static void cancelAlarm(Context context) {
         Intent intent = new Intent(context, KeepAliveReceiver.class);
@@ -38,35 +44,49 @@ public class KeepAliveReceiver extends BroadcastReceiver {
     @Override
     public void onReceive(Context context, Intent rIntent) {
         PowerManager pm = (PowerManager) context.getSystemService(Context.POWER_SERVICE);
-        PowerManager.WakeLock wl = pm.newWakeLock(PowerManager.PARTIAL_WAKE_LOCK, "");
+        PowerManager.WakeLock wl = pm.newWakeLock(PowerManager.PARTIAL_WAKE_LOCK, "AndroidAPS:KeepAliveReciever");
         wl.acquire();
 
         LocalAlertUtils.shortenSnoozeInterval();
         LocalAlertUtils.checkStaleBGAlert();
         checkPump();
 
-        log.debug("KeepAlive received");
+        if (L.isEnabled(L.CORE))
+            log.debug("KeepAlive received");
         wl.release();
     }
 
     private void checkPump() {
-        final PumpInterface pump = ConfigBuilderPlugin.getActivePump();
-        final Profile profile = MainApp.getConfigBuilder().getProfile();
-        if (pump != null && profile != null && profile.getBasal() != null) {
-            Date lastConnection = pump.lastDataTime();
-            boolean isStatusOutdated = lastConnection.getTime() + STATUS_UPDATE_FREQUENCY < System.currentTimeMillis();
+        final PumpInterface pump = ConfigBuilderPlugin.getPlugin().getActivePump();
+        final Profile profile = ProfileFunctions.getInstance().getProfile();
+        if (pump != null && profile != null) {
+            long lastConnection = pump.lastDataTime();
+            boolean isStatusOutdated = lastConnection + STATUS_UPDATE_FREQUENCY < System.currentTimeMillis();
             boolean isBasalOutdated = Math.abs(profile.getBasal() - pump.getBaseBasalRate()) > pump.getPumpDescription().basalStep;
 
-            LocalAlertUtils.checkPumpUnreachableAlarm(lastConnection, isStatusOutdated);
+            if (L.isEnabled(L.CORE))
+                log.debug("Last connection: " + DateUtil.dateAndTimeString(lastConnection));
+            // sometimes keepalive broadcast stops
+            // as as workaround test if readStatus was requested before an alarm is generated
+            if (lastReadStatus != 0 && lastReadStatus > System.currentTimeMillis() - T.mins(5).msecs()) {
+                LocalAlertUtils.checkPumpUnreachableAlarm(lastConnection, isStatusOutdated);
+            }
 
-            if (!pump.isThisProfileSet(profile)) {
-                MainApp.getConfigBuilder().getCommandQueue().setProfile(profile, null);
+            if (!pump.isThisProfileSet(profile) && !ConfigBuilderPlugin.getPlugin().getCommandQueue().isRunning(Command.CommandType.BASALPROFILE)) {
+                MainApp.bus().post(new EventProfileNeedsUpdate());
             } else if (isStatusOutdated && !pump.isBusy()) {
-                MainApp.getConfigBuilder().getCommandQueue().readStatus("KeepAlive. Status outdated.", null);
+                lastReadStatus = System.currentTimeMillis();
+                ConfigBuilderPlugin.getPlugin().getCommandQueue().readStatus("KeepAlive. Status outdated.", null);
             } else if (isBasalOutdated && !pump.isBusy()) {
-                MainApp.getConfigBuilder().getCommandQueue().readStatus("KeepAlive. Basal outdated.", null);
+                lastReadStatus = System.currentTimeMillis();
+                ConfigBuilderPlugin.getPlugin().getCommandQueue().readStatus("KeepAlive. Basal outdated.", null);
             }
         }
+        if (lastRun != 0 && System.currentTimeMillis() - lastRun > T.mins(10).msecs()) {
+            log.error("KeepAlive fail");
+            FabricPrivacy.getInstance().logCustom("KeepAliveFail");
+        }
+        lastRun = System.currentTimeMillis();
     }
 
     //called by MainApp at first app start
