@@ -11,14 +11,15 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import info.nightscout.androidaps.MainApp;
-import info.nightscout.androidaps.Services.Intents;
-import info.nightscout.androidaps.plugins.NSClientInternal.data.NSAlarm;
-import info.nightscout.androidaps.plugins.Overview.notifications.Notification;
-import info.nightscout.androidaps.plugins.Overview.events.EventDismissNotification;
-import info.nightscout.androidaps.plugins.Overview.events.EventNewNotification;
+import info.nightscout.androidaps.logging.L;
+import info.nightscout.androidaps.plugins.general.nsclient.data.NSAlarm;
+import info.nightscout.androidaps.plugins.general.overview.events.EventDismissNotification;
+import info.nightscout.androidaps.plugins.general.overview.events.EventNewNotification;
+import info.nightscout.androidaps.plugins.general.overview.notifications.Notification;
+import info.nightscout.androidaps.services.Intents;
 
 public class NSAlarmReceiver extends BroadcastReceiver {
-    private static Logger log = LoggerFactory.getLogger(NSAlarmReceiver.class);
+    private static Logger log = LoggerFactory.getLogger(L.CORE);
 
     @Override
     public void onReceive(Context context, Intent intent) {
