@@ -1,5 +1,7 @@
 package info.nightscout.androidaps.data;
 
+import java.util.Objects;
+
 /**
  * Created by emmablack on 1/7/15.
  */
@@ -8,12 +10,18 @@ public class BgWatchData implements Comparable<BgWatchData>{
     public double high;
     public double low;
     public long timestamp;
+    public int color;
 
-    public BgWatchData(double aSgv, double aHigh, double aLow, long aTimestamp) {
+    public BgWatchData(double aSgv, double aHigh, double aLow, long aTimestamp, int aColor) {
         this.sgv = aSgv;
         this.high = aHigh;
         this.low = aLow;
         this.timestamp = aTimestamp;
+        this.color = aColor;
+    }
+
+    public BgWatchData(){
+
     }
 
     @Override
@@ -21,12 +29,14 @@ public class BgWatchData implements Comparable<BgWatchData>{
         if(! (that instanceof BgWatchData)){
             return  false;
         }
+        if (this.color != ((BgWatchData) that).color)
+            return false;
         return this.timestamp == ((BgWatchData) that).timestamp;
     }
 
     @Override
-    public int hashCode(){
-        return (int) (timestamp%Integer.MAX_VALUE);
+    public int hashCode() {
+        return Objects.hash(timestamp, color);
     }
 
     @Override
