@@ -28,17 +28,16 @@ import android.view.WindowManager;
 import android.widget.TextView;
 
 import com.google.android.gms.wearable.DataMap;
-import com.ustwo.clockwise.wearable.WatchFace;
 import com.ustwo.clockwise.common.WatchFaceTime;
+import com.ustwo.clockwise.wearable.WatchFace;
 
 import java.util.ArrayList;
 import java.util.Calendar;
-import java.util.Date;
 import java.util.HashSet;
 import java.util.TreeSet;
 
-import info.nightscout.androidaps.data.BgWatchData;
 import info.nightscout.androidaps.R;
+import info.nightscout.androidaps.data.BgWatchData;
 import info.nightscout.androidaps.interaction.menus.MainMenuActivity;
 
 
@@ -77,9 +76,8 @@ public class CircleWatchface extends WatchFace implements SharedPreferences.OnSh
     private String direction = "";
     private String delta = "";
     private String avgDelta = "";
-    public TreeSet<BgWatchData> bgDataList = new TreeSet<BgWatchData>();
+    public TreeSet<BgWatchData> bgDataList = new TreeSet<>();
 
-    private View layoutView;
     private int specW;
     private int specH;
     private View myLayout;
@@ -94,7 +92,7 @@ public class CircleWatchface extends WatchFace implements SharedPreferences.OnSh
         super.onCreate();
 
         PowerManager powerManager = (PowerManager) getSystemService(POWER_SERVICE);
-        PowerManager.WakeLock wakeLock = powerManager.newWakeLock(PowerManager.PARTIAL_WAKE_LOCK, "CreateWakelock");
+        PowerManager.WakeLock wakeLock = powerManager.newWakeLock(PowerManager.PARTIAL_WAKE_LOCK, "AndroidAPS:CircleWatchface");
         wakeLock.acquire(30000);
 
         Display display = ((WindowManager) getSystemService(Context.WINDOW_SERVICE))
@@ -151,7 +149,7 @@ public class CircleWatchface extends WatchFace implements SharedPreferences.OnSh
 
         // prepare fields
 
-        TextView textView = null;
+        TextView textView;
         mSgv = (TextView) myLayout.findViewById(R.id.sgvString);
         textView = (TextView) myLayout.findViewById(R.id.sgvString);
         if (sharedPrefs.getBoolean("showBG", true)) {
@@ -201,7 +199,7 @@ public class CircleWatchface extends WatchFace implements SharedPreferences.OnSh
             } else {
                 textView.setTextSize(TypedValue.COMPLEX_UNIT_SP, 18);
             }
-            if(sharedPrefs.getBoolean("showAvgDelta", true)){
+            if (sharedPrefs.getBoolean("showAvgDelta", true)) {
                 textView.append("  " + getAvgDelta());
             }
 
@@ -298,10 +296,8 @@ public class CircleWatchface extends WatchFace implements SharedPreferences.OnSh
         removePaint.setAntiAlias(true);
         removePaint.setColor(getBackgroundColor());
 
-        ;
-
-        rect = new RectF(PADDING, PADDING, (float) (displaySize.x - PADDING), (float) (displaySize.y - PADDING));
-        rectDelete = new RectF(PADDING - CIRCLE_WIDTH / 2, PADDING - CIRCLE_WIDTH / 2, (float) (displaySize.x - PADDING + CIRCLE_WIDTH / 2), (float) (displaySize.y - PADDING + CIRCLE_WIDTH / 2));
+        rect = new RectF(PADDING, PADDING, displaySize.x - PADDING, displaySize.y - PADDING);
+        rectDelete = new RectF(PADDING - CIRCLE_WIDTH / 2, PADDING - CIRCLE_WIDTH / 2, displaySize.x - PADDING + CIRCLE_WIDTH / 2, displaySize.y - PADDING + CIRCLE_WIDTH / 2);
         overlapping = ALWAYS_HIGHLIGT_SMALL || areOverlapping(angleSMALL, angleSMALL + SMALL_HAND_WIDTH + NEAR, angleBig, angleBig + BIG_HAND_WIDTH + NEAR);
         Log.d("CircleWatchface", "end prepareDrawTime");
 
@@ -326,11 +322,11 @@ public class CircleWatchface extends WatchFace implements SharedPreferences.OnSh
     protected void onTimeChanged(WatchFaceTime oldTime, WatchFaceTime newTime) {
         if (oldTime.hasMinuteChanged(newTime)) {
             PowerManager powerManager = (PowerManager) getSystemService(POWER_SERVICE);
-            PowerManager.WakeLock wakeLock = powerManager.newWakeLock(PowerManager.PARTIAL_WAKE_LOCK, "TimeChangedWakelock");
+            PowerManager.WakeLock wakeLock = powerManager.newWakeLock(PowerManager.PARTIAL_WAKE_LOCK, "AndroidAPS:CircleWatchface_onTimeChanged");
             wakeLock.acquire(30000);
             /*Preparing the layout just on every minute tick:
-            *  - hopefully better battery life
-            *  - drawback: might update the minutes since last reading up endTime one minute late*/
+             *  - hopefully better battery life
+             *  - drawback: might update the minutes since last reading up endTime one minute late*/
             prepareLayout();
             prepareDrawTime();
             invalidate();  //redraw the time
@@ -476,6 +472,7 @@ public class CircleWatchface extends WatchFace implements SharedPreferences.OnSh
     private String getAvgDelta() {
         return avgDelta;
     }
+
     private void setAvgDelta(String avgDelta) {
         this.avgDelta = avgDelta;
     }
@@ -522,11 +519,10 @@ public class CircleWatchface extends WatchFace implements SharedPreferences.OnSh
         @Override
         public void onReceive(Context context, Intent intent) {
             PowerManager powerManager = (PowerManager) getSystemService(POWER_SERVICE);
-            PowerManager.WakeLock wakeLock = powerManager.newWakeLock(PowerManager.PARTIAL_WAKE_LOCK,
-                    "MyWakelockTag");
+            PowerManager.WakeLock wakeLock = powerManager.newWakeLock(PowerManager.PARTIAL_WAKE_LOCK, "AndroidAPS:MessageReceiver");
             wakeLock.acquire(30000);
             Bundle bundle = intent.getBundleExtra("data");
-            if (bundle!= null) {
+            if (bundle != null) {
                 DataMap dataMap = DataMap.fromBundle(bundle);
                 setSgvLevel((int) dataMap.getLong("sgvLevel"));
                 Log.d("CircleWatchface", "sgv level : " + getSgvLevel());
@@ -565,7 +561,7 @@ public class CircleWatchface extends WatchFace implements SharedPreferences.OnSh
 
     public synchronized void addToWatchSet(DataMap dataMap) {
 
-        if(!sharedPrefs.getBoolean("showRingHistory", false)){
+        if (!sharedPrefs.getBoolean("showRingHistory", false)) {
             bgDataList.clear();
             return;
         }
@@ -577,7 +573,8 @@ public class CircleWatchface extends WatchFace implements SharedPreferences.OnSh
             double high = dataMap.getDouble("high");
             double low = dataMap.getDouble("low");
             long timestamp = dataMap.getLong("timestamp");
-            bgDataList.add(new BgWatchData(sgv, high, low, timestamp));
+            int color = dataMap.getInt("color", 0);
+            bgDataList.add(new BgWatchData(sgv, high, low, timestamp, color));
         } else if (!sharedPrefs.getBoolean("animation", false)) {
             // don't load history at once if animations are set (less resource consumption)
             Log.d("addToWatchSet", "entries.size(): " + entries.size());
@@ -587,12 +584,13 @@ public class CircleWatchface extends WatchFace implements SharedPreferences.OnSh
                 double high = entry.getDouble("high");
                 double low = entry.getDouble("low");
                 long timestamp = entry.getLong("timestamp");
-                bgDataList.add(new BgWatchData(sgv, high, low, timestamp));
+                int color = entry.getInt("color", 0);
+                bgDataList.add(new BgWatchData(sgv, high, low, timestamp, color));
             }
         } else
 
             Log.d("addToWatchSet", "start removing bgDataList.size(): " + bgDataList.size());
-        HashSet removeSet = new HashSet();
+        HashSet<BgWatchData> removeSet = new HashSet<>();
         double threshold = (System.currentTimeMillis() - (1000 * 60 * 5 * holdInMemory()));
         for (BgWatchData data : bgDataList) {
             if (data.timestamp < threshold) {
@@ -708,14 +706,14 @@ public class CircleWatchface extends WatchFace implements SharedPreferences.OnSh
     @Override
     protected void onTapCommand(int tapType, int x, int y, long eventTime) {
 
-        int extra = mSgv!=null?(mSgv.getRight() - mSgv.getLeft())/2:0;
+        int extra = mSgv != null ? (mSgv.getRight() - mSgv.getLeft()) / 2 : 0;
 
-        if (tapType == TAP_TYPE_TAP&&
-                x + extra >=mSgv.getLeft() &&
-                x - extra <= mSgv.getRight()&&
+        if (tapType == TAP_TYPE_TAP &&
+                x + extra >= mSgv.getLeft() &&
+                x - extra <= mSgv.getRight() &&
                 y >= mSgv.getTop() &&
-                y <= mSgv.getBottom()){
-            if (eventTime - sgvTapTime < 800){
+                y <= mSgv.getBottom()) {
+            if (eventTime - sgvTapTime < 800) {
                 Intent intent = new Intent(this, MainMenuActivity.class);
                 intent.addFlags(Intent.FLAG_ACTIVITY_NEW_TASK);
                 startActivity(intent);
@@ -725,7 +723,7 @@ public class CircleWatchface extends WatchFace implements SharedPreferences.OnSh
     }
 
     @Override
-    protected WatchFaceStyle getWatchFaceStyle(){
+    protected WatchFaceStyle getWatchFaceStyle() {
         return new WatchFaceStyle.Builder(this).setAcceptsTapEvents(true).build();
     }
 
